refactor(management): decompose monitoring center
