+++ cliproxyapi-pro-management/overlay/src/features/monitoring/monitoringOptions.ts
@@ -1,0 +1,23 @@
+import type { MonitoringTimeRange } from './hooks/useMonitoringData';
+import type { AccountSortMetric, RankingMetric } from './monitoringAnalytics';
+
+export const TIME_RANGE_OPTIONS: Array<{ value: MonitoringTimeRange; labelKey: string }> = [
+  { value: 'today', labelKey: 'monitoring.range_today' },
+  { value: '7d', labelKey: 'monitoring.range_7d' },
+  { value: '14d', labelKey: 'monitoring.range_14d' },
+  { value: '30d', labelKey: 'monitoring.range_30d' },
+  { value: 'all', labelKey: 'monitoring.range_all' },
+];
+
+export const RANKING_METRIC_OPTIONS: Array<{ value: RankingMetric; labelKey: string }> = [
+  { value: 'requests', labelKey: 'monitoring.ranking_metric_requests' },
+  { value: 'tokens', labelKey: 'monitoring.ranking_metric_tokens' },
+  { value: 'cost', labelKey: 'monitoring.ranking_metric_cost' },
+];
+
+export const ACCOUNT_SORT_OPTIONS: Array<{ value: AccountSortMetric; labelKey: string }> = [
+  { value: 'recent', labelKey: 'monitoring.account_sort_recent' },
+  { value: 'requests', labelKey: 'monitoring.ranking_metric_requests' },
+  { value: 'tokens', labelKey: 'monitoring.ranking_metric_tokens' },
+  { value: 'cost', labelKey: 'monitoring.ranking_metric_cost' },
+];
