+++ cliproxyapi-pro-management/overlay/src/pages/MonitoringCenterPage.tsx
@@ -1,5 +1,4 @@
-import { useCallback, useDeferredValue, useEffect, useId, useMemo, useRef, useState, type ChangeEvent, type CSSProperties, type DragEvent, type MouseEvent as ReactMouseEvent, type ReactNode } from 'react';
-import { createPortal } from 'react-dom';
+import { useCallback, useDeferredValue, useEffect, useMemo, useRef, useState, type ChangeEvent, type CSSProperties, type DragEvent, type MouseEvent as ReactMouseEvent, type ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { TFunction } from 'i18next';
 import { Button } from '@/components/ui/Button';
@@ -10,7 +9,6 @@ import { Select } from '@/components/ui/Select';
 import {
   IconChevronDown,
   IconChevronUp,
-  IconInfo,
   IconRefreshCw,
   IconSearch,
   IconSlidersHorizontal,
@@ -18,8 +16,6 @@ import {
 } from '@/components/ui/icons';
 import {
   buildAccountRowsByAccount,
-  buildDayLabel,
-  buildHourLabel,
   buildLocalDayKey,
   formatShortDateTime,
   getRangeStartMs,
@@ -27,7 +23,6 @@ import {
   useMonitoringData,
   type MonitoringAccountRow,
   type MonitoringEventRow,
-  type MonitoringSummary,
   type MonitoringStatusTone,
   type MonitoringTimeRange,
 } from '@/features/monitoring/hooks/useMonitoringData';
@@ -35,6 +30,71 @@ import { REALTIME_LOG_PAGE_SIZE, useRealtimeLogData } from '@/features/monitorin
 import { useUsageData, type UsageEventPageFilters, type UsagePayload } from '@/features/monitoring/hooks/useUsageData';
 import { useUsageAggregates, type UsageAggregateBucket } from '@/features/monitoring/hooks/useUsageAggregates';
 import { findMonitoringAuthIndexes } from '@/features/monitoring/monitoringAuthSearch';
+import {
+  buildAccountStatusData,
+  buildAccountStatusRange,
+  type AccountStatusData,
+} from '@/features/monitoring/accountHealth';
+import { MonitoringHealthStatusBar } from '@/features/monitoring/components/MonitoringHealthStatusBar';
+import {
+  RealtimeCostCell,
+} from '@/features/monitoring/components/RealtimeCostCell';
+import {
+  ApiKeyRankingPanel,
+  ModelStatsPanel,
+  TokenDistributionPanel,
+  TopUsageStats,
+  UsageTrendHeader,
+  UsageTrendPanel,
+  type UsageMetricCard,
+} from '@/features/monitoring/components/UsageAnalyticsPanels';
+import {
+  buildAggregateSummary,
+  buildServerAccountRows,
+} from '@/features/monitoring/monitoringAggregates';
+import {
+  addMonitoringSummaryRow,
+  buildServerUsageTrendAnalytics,
+  buildUsageTrendAnalytics,
+  buildUsageTrendRangeLabel,
+  createMonitoringSummaryAccumulator,
+  finalizeMonitoringSummary,
+  formatPercent,
+  getAccountHealthTone,
+  getAccountSortValue,
+  getRankingMetricValue,
+  type AccountHealthTone,
+  type AccountSortMetric,
+  type RankingMetric,
+} from '@/features/monitoring/monitoringAnalytics';
+import {
+  ACCOUNT_SORT_OPTIONS,
+  TIME_RANGE_OPTIONS,
+} from '@/features/monitoring/monitoringOptions';
+import {
+  REALTIME_LOG_COLUMN_DEFAULT_WIDTHS,
+  clampRealtimeLogColumnWidth,
+  createDefaultRealtimeLogColumns,
+  isRealtimeLogColumnKey,
+  loadRealtimeLogColumns,
+  loadRealtimeLogFollowEnabled,
+  normalizeRealtimeLogColumns,
+  saveRealtimeLogColumns,
+  saveRealtimeLogFollowEnabled,
+  type RealtimeLogColumnKey,
+  type RealtimeLogColumnPreference,
+} from '@/features/monitoring/realtimeLogPreferences';
+import {
+  buildRealtimeDiagnosticClipboardText,
+  buildRealtimeLogPageRows,
+  buildRealtimeMetaText,
+  buildRealtimeStatusLabel,
+  compactRealtimeErrorMessage,
+  getClientPaginationRange,
+  translateRealtimeErrorCategory,
+  translateRealtimeErrorText,
+  type RealtimeLogRow,
+} from '@/features/monitoring/realtimeLogPresentation';
 import { hasUsageBackupManifest } from '@/features/monitoring/usageBackup';
 import { useHeaderRefresh } from '@/hooks/useHeaderRefresh';
 import { apiClient } from '@/services/api/client';
@@ -43,7 +103,6 @@ import type { AuthFileItem } from '@/types';
 import { maskSensitiveText } from '@/utils/format';
 import { getStatusFromError, isAntigravityFile, isClaudeFile, isCodexFile, isKimiFile, isXaiFile } from '@/utils/quota';
 import {
-  calculateCost,
   deleteModelPriceRule,
   formatCompactNumber,
   formatDurationMs,
@@ -55,7 +114,6 @@ import {
   recalculateModelPriceHistory,
   saveModelPriceRule,
   syncModelPricesFromModelsDev,
-  type ModelPrice,
   type ModelPriceRule,
   type ModelPriceSyncChangeAction,
   type ModelPriceSyncResult,
@@ -78,71 +136,10 @@ import quotaStyles from '@/pages/QuotaPage.module.scss';
 import { quotaPersistenceMiddleware } from '@/extensions/quota/persistenceMiddleware';
 import styles from './MonitoringCenterPage.module.scss';
 
-const TIME_RANGE_OPTIONS: Array<{ value: MonitoringTimeRange; labelKey: string }> = [
-  { value: 'today', labelKey: 'monitoring.range_today' },
-  { value: '7d', labelKey: 'monitoring.range_7d' },
-  { value: '14d', labelKey: 'monitoring.range_14d' },
-  { value: '30d', labelKey: 'monitoring.range_30d' },
-  { value: 'all', labelKey: 'monitoring.range_all' },
-];
-
 type StatusFilter = 'all' | 'success' | 'failed';
 
-type UsageMetricCard = {
-  key: string;
-  title: string;
-  label: string;
-  value: ReactNode;
-  accent: 'blue' | 'purple' | 'green' | 'amber';
-  footer: Array<{ label: string; value: ReactNode }>;
-};
-
-type RankingMetric = 'requests' | 'tokens' | 'cost';
-type AccountSortMetric = 'recent' | RankingMetric;
-
-const RANKING_METRIC_OPTIONS: Array<{ value: RankingMetric; labelKey: string }> = [
-  { value: 'requests', labelKey: 'monitoring.ranking_metric_requests' },
-  { value: 'tokens', labelKey: 'monitoring.ranking_metric_tokens' },
-  { value: 'cost', labelKey: 'monitoring.ranking_metric_cost' },
-];
-
-const ACCOUNT_SORT_OPTIONS: Array<{ value: AccountSortMetric; labelKey: string }> = [
-  { value: 'recent', labelKey: 'monitoring.account_sort_recent' },
-  { value: 'requests', labelKey: 'monitoring.ranking_metric_requests' },
-  { value: 'tokens', labelKey: 'monitoring.ranking_metric_tokens' },
-  { value: 'cost', labelKey: 'monitoring.ranking_metric_cost' },
-];
-
-const ACCOUNT_STATUS_BLOCK_COUNT = 20;
-const ACCOUNT_STATUS_BLOCK_DURATION_MS = 10 * 60 * 1000;
 const ACCOUNT_STATS_ANALYTICS_ROW_LIMIT = 6000;
-const REALTIME_LOG_ENRICH_LIMIT = REALTIME_LOG_PAGE_SIZE;
 const ACCOUNT_QUOTA_REQUEST_CONCURRENCY = 4;
-const REALTIME_LOG_COLUMNS_STORAGE_KEY = 'cli-proxy-realtime-log-columns-v2';
-const REALTIME_LOG_FOLLOW_STORAGE_KEY = 'cli-proxy-realtime-log-follow-v1';
-const REALTIME_LOG_COLUMN_KEYS = [
-  'type',
-  'model',
-  'reasoningEffort',
-  'stream',
-  'apiKey',
-  'recent',
-  'status',
-  'successRate',
-  'calls',
-  'ttft',
-  'latency',
-  'tokens',
-  'cacheRead',
-  'cost',
-  'time',
-] as const;
-type RealtimeLogColumnKey = typeof REALTIME_LOG_COLUMN_KEYS[number];
-type RealtimeLogColumnPreference = {
-  key: RealtimeLogColumnKey;
-  visible: boolean;
-  width?: number;
-};
 type RealtimeLogColumnDefinition = {
   key: RealtimeLogColumnKey;
   label: string;
@@ -152,194 +149,11 @@ type RealtimeLogColumnDefinition = {
   render: (row: RealtimeLogRow) => ReactNode;
   width: number;
 };
-const REALTIME_LOG_COLUMN_DEFAULT_WIDTHS: Record<RealtimeLogColumnKey, number> = {
-  type: 170,
-  model: 230,
-  reasoningEffort: 116,
-  stream: 108,
-  apiKey: 145,
-  recent: 86,
-  status: 180,
-  successRate: 86,
-  calls: 76,
-  ttft: 92,
-  latency: 96,
-  tokens: 196,
-  cacheRead: 126,
-  cost: 132,
-  time: 164,
-};
-const REALTIME_LOG_COLUMN_MIN_WIDTHS: Record<RealtimeLogColumnKey, number> = {
-  type: 96,
-  model: 132,
-  reasoningEffort: 96,
-  stream: 92,
-  apiKey: 104,
-  recent: 76,
-  status: 120,
-  successRate: 76,
-  calls: 68,
-  ttft: 76,
-  latency: 76,
-  tokens: 164,
-  cacheRead: 108,
-  cost: 112,
-  time: 116,
-};
-const REALTIME_LOG_COLUMN_MAX_WIDTH = 420;
-const REALTIME_LOG_COLUMN_MAX_WIDTHS: Partial<Record<RealtimeLogColumnKey, number>> = {
-  type: 240,
-};
-const REALTIME_LOG_COLUMN_KEY_SET = new Set<RealtimeLogColumnKey>(REALTIME_LOG_COLUMN_KEYS);
-const createDefaultRealtimeLogColumns = (): RealtimeLogColumnPreference[] => (
-  REALTIME_LOG_COLUMN_KEYS.map((key) => ({ key, visible: true }))
-);
-const REALTIME_LOG_DEFAULT_COLUMNS = createDefaultRealtimeLogColumns();
-const ACCOUNT_STATUS_COLOR_STOPS = [
-  { r: 239, g: 68, b: 68 },
-  { r: 250, g: 204, b: 21 },
-  { r: 34, g: 197, b: 94 },
-] as const;
-
-const getAccountStatusColor = (rate: number) => {
-  const value = Math.max(0, Math.min(1, rate));
-  const segment = value < 0.5 ? 0 : 1;
-  const localValue = segment === 0 ? value * 2 : (value - 0.5) * 2;
-  const from = ACCOUNT_STATUS_COLOR_STOPS[segment];
-  const to = ACCOUNT_STATUS_COLOR_STOPS[segment + 1];
-  const r = Math.round(from.r + (to.r - from.r) * localValue);
-  const g = Math.round(from.g + (to.g - from.g) * localValue);
-  const b = Math.round(from.b + (to.b - from.b) * localValue);
-  return `rgb(${r}, ${g}, ${b})`;
-};
-
-const formatStatusRate = (rate: number) => {
-  const rounded = rate.toFixed(1);
-  return `${rounded.endsWith('.0') ? rounded.slice(0, -2) : rounded}%`;
-};
-
 const formatTokenCount = (value: number) => Math.max(0, Math.round(Number(value) || 0)).toLocaleString();
 
 const getCacheHitRate = (row: Pick<MonitoringEventRow, 'inputTokens' | 'cachedTokens'>): number | null => (
   row.inputTokens > 0 ? Math.min(Math.max(row.cachedTokens / row.inputTokens, 0), 1) : null
 );
-
-const isRealtimeLogColumnKey = (value: unknown): value is RealtimeLogColumnKey => (
-  typeof value === 'string' && REALTIME_LOG_COLUMN_KEY_SET.has(value as RealtimeLogColumnKey)
-);
-
-const clampRealtimeLogColumnWidth = (key: RealtimeLogColumnKey, width: unknown) => {
-  const numericWidth = typeof width === 'number' && Number.isFinite(width)
-    ? width
-    : REALTIME_LOG_COLUMN_DEFAULT_WIDTHS[key];
-  const maxWidth = REALTIME_LOG_COLUMN_MAX_WIDTHS[key] ?? REALTIME_LOG_COLUMN_MAX_WIDTH;
-  return Math.min(maxWidth, Math.max(REALTIME_LOG_COLUMN_MIN_WIDTHS[key], Math.round(numericWidth)));
-};
-
-const normalizeRealtimeLogColumnWidth = (key: RealtimeLogColumnKey, width: unknown) => (
-  typeof width === 'number' && Number.isFinite(width)
-    ? clampRealtimeLogColumnWidth(key, width)
-    : undefined
-);
-
-const normalizeRealtimeLogColumns = (value: unknown): RealtimeLogColumnPreference[] => {
-  const next: RealtimeLogColumnPreference[] = [];
-  const seen = new Set<RealtimeLogColumnKey>();
-
-  if (Array.isArray(value)) {
-    value.forEach((item) => {
-      if (!item || typeof item !== 'object') return;
-      const key = (item as { key?: unknown }).key;
-      if (key === 'usage') {
-        const visible = (item as { visible?: unknown }).visible !== false;
-        (['tokens', 'cacheRead'] as const).forEach((replacementKey) => {
-          if (seen.has(replacementKey)) return;
-          next.push({ key: replacementKey, visible });
-          seen.add(replacementKey);
-        });
-        return;
-      }
-      if (!isRealtimeLogColumnKey(key) || seen.has(key)) return;
-      next.push({
-        key,
-        visible: (item as { visible?: unknown }).visible !== false,
-        width: normalizeRealtimeLogColumnWidth(key, (item as { width?: unknown }).width),
-      });
-      seen.add(key);
-    });
-  }
-
-  const shouldMigrateReasoningEffort = next.length > 0 && !seen.has('reasoningEffort');
-  const shouldMigrateStream = next.length > 0 && !seen.has('stream');
-
-  REALTIME_LOG_DEFAULT_COLUMNS.forEach((item) => {
-    if (!seen.has(item.key)) {
-      next.push({ ...item });
-    }
-  });
-
-  if (shouldMigrateReasoningEffort) {
-    const reasoningEffortIndex = next.findIndex((item) => item.key === 'reasoningEffort');
-    const modelIndex = next.findIndex((item) => item.key === 'model');
-    if (reasoningEffortIndex >= 0 && modelIndex >= 0) {
-      const [reasoningEffortColumn] = next.splice(reasoningEffortIndex, 1);
-      const migratedModelIndex = next.findIndex((item) => item.key === 'model');
-      next.splice(migratedModelIndex + 1, 0, reasoningEffortColumn);
-    }
-  }
-
-  if (shouldMigrateStream) {
-    const streamIndex = next.findIndex((item) => item.key === 'stream');
-    const reasoningEffortIndex = next.findIndex((item) => item.key === 'reasoningEffort');
-    if (streamIndex >= 0 && reasoningEffortIndex >= 0) {
-      const [streamColumn] = next.splice(streamIndex, 1);
-      const migratedReasoningEffortIndex = next.findIndex((item) => item.key === 'reasoningEffort');
-      next.splice(migratedReasoningEffortIndex + 1, 0, streamColumn);
-    }
-  }
-
-  const timeColumn = next.find((item) => item.key === 'time');
-  const ordered = timeColumn ? [...next.filter((item) => item.key !== 'time'), timeColumn] : next;
-  return ordered.some((item) => item.visible) ? ordered : createDefaultRealtimeLogColumns();
-};
-
-const loadRealtimeLogColumns = () => {
-  if (typeof window === 'undefined') {
-    return createDefaultRealtimeLogColumns();
-  }
-  try {
-    return normalizeRealtimeLogColumns(JSON.parse(window.localStorage.getItem(REALTIME_LOG_COLUMNS_STORAGE_KEY) || 'null'));
-  } catch {
-    return createDefaultRealtimeLogColumns();
-  }
-};
-
-const saveRealtimeLogColumns = (columns: RealtimeLogColumnPreference[]) => {
-  if (typeof window === 'undefined') return;
-  try {
-    window.localStorage.setItem(REALTIME_LOG_COLUMNS_STORAGE_KEY, JSON.stringify(columns));
-  } catch {
-    // Column preferences are convenience state; storage failures should not break logs.
-  }
-};
-
-const loadRealtimeLogFollowEnabled = () => {
-  if (typeof window === 'undefined') return true;
-  try {
-    return window.localStorage.getItem(REALTIME_LOG_FOLLOW_STORAGE_KEY) !== 'false';
-  } catch {
-    return true;
-  }
-};
-
-const saveRealtimeLogFollowEnabled = (enabled: boolean) => {
-  if (typeof window === 'undefined') return;
-  try {
-    window.localStorage.setItem(REALTIME_LOG_FOLLOW_STORAGE_KEY, String(enabled));
-  } catch {
-    // Live-follow preference is optional; storage failures should not break logs.
-  }
-};
 
 const getRealtimeLogColumnContentTexts = (key: RealtimeLogColumnKey, row: RealtimeLogRow) => {
   switch (key) {
@@ -407,198 +221,7 @@ const estimateRealtimeLogHeaderWidth = (key: RealtimeLogColumnKey, label: string
   return clampRealtimeLogColumnWidth(key, textWidth + 42);
 };
 
-const formatStatusWindowLabel = (startTime: number, endTime: number, locale: string) => {
-  const start = new Date(startTime);
-  const end = new Date(endTime);
-  const sameDay = start.toDateString() === end.toDateString();
-  const dateOptions: Intl.DateTimeFormatOptions = { month: 'numeric', day: 'numeric' };
-  const timeOptions: Intl.DateTimeFormatOptions = { hour: '2-digit', minute: '2-digit' };
-  const startDateLabel = start.toLocaleDateString(locale, dateOptions);
-  const endDateLabel = end.toLocaleDateString(locale, dateOptions);
-  const startTimeLabel = start.toLocaleTimeString(locale, timeOptions);
-  const endTimeLabel = end.toLocaleTimeString(locale, timeOptions);
-
-  return sameDay
-    ? `${startDateLabel} ${startTimeLabel} - ${endTimeLabel}`
-    : `${startDateLabel} ${startTimeLabel} - ${endDateLabel} ${endTimeLabel}`;
-};
-
-const buildAccountStatusBlockAriaLabel = ({
-  detail,
-  timeRangeLabel,
-  successRateValue,
-  copy,
-}: {
-  detail: AccountStatusBlockDetail;
-  timeRangeLabel: string;
-  successRateValue: string;
-  copy: {
-    successLabel: string;
-    failureLabel: string;
-    noRequestsLabel: string;
-    successRateLabel: string;
-  };
-}) => {
-  const total = detail.success + detail.failure;
-  if (total === 0) {
-    return `${timeRangeLabel}，${copy.noRequestsLabel}`;
-  }
-  return `${timeRangeLabel}，${copy.successLabel} ${detail.success}，${copy.failureLabel} ${detail.failure}，${copy.successRateLabel} ${successRateValue}`;
-};
-
-const getNextAccountStatusBlockIndex = (currentIndex: number, key: string, count: number) => {
-  if (count <= 0) return null;
-  if (key === 'ArrowRight' || key === 'ArrowDown') return Math.min(currentIndex + 1, count - 1);
-  if (key === 'ArrowLeft' || key === 'ArrowUp') return Math.max(currentIndex - 1, 0);
-  if (key === 'Home') return 0;
-  if (key === 'End') return count - 1;
-  return null;
-};
-
 const formatAccountOverviewScopeText = (rangeLabel: string, t: TFunction) => t('monitoring.account_scope_text', { range: rangeLabel });
-
-const buildAccountStatusRange = (rows: MonitoringAccountRow[], range: MonitoringTimeRange, nowMs = Date.now()): AccountStatusRange => {
-  if (range !== 'all') {
-    return {
-      startTime: getRangeStartMs(range, nowMs),
-      endTime: nowMs,
-    };
-  }
-
-  let minTimestamp = Number.POSITIVE_INFINITY;
-  rows.forEach((row) => {
-    row.rows?.forEach((event) => {
-      minTimestamp = Math.min(minTimestamp, event.timestampMs);
-    });
-  });
-  return {
-    startTime: Number.isFinite(minTimestamp) ? minTimestamp : nowMs - ACCOUNT_STATUS_BLOCK_COUNT * ACCOUNT_STATUS_BLOCK_DURATION_MS,
-    endTime: nowMs,
-  };
-};
-
-const buildAccountStatusData = (rows: MonitoringEventRow[], range: AccountStatusRange): AccountStatusData => {
-  const duration = Math.max(range.endTime - range.startTime, ACCOUNT_STATUS_BLOCK_DURATION_MS);
-  const blockDuration = duration / ACCOUNT_STATUS_BLOCK_COUNT;
-  const blockDetails = Array.from({ length: ACCOUNT_STATUS_BLOCK_COUNT }, (_, index) => ({
-    success: 0,
-    failure: 0,
-    rate: -1,
-    startTime: range.startTime + index * blockDuration,
-    endTime: index === ACCOUNT_STATUS_BLOCK_COUNT - 1 ? range.endTime : range.startTime + (index + 1) * blockDuration,
-  }));
-  let totalSuccess = 0;
-  let totalFailure = 0;
-
-  rows.forEach((row) => {
-    if (row.timestampMs < range.startTime || row.timestampMs > range.endTime) return;
-    const index = Math.min(
-      ACCOUNT_STATUS_BLOCK_COUNT - 1,
-      Math.max(0, Math.floor((row.timestampMs - range.startTime) / blockDuration))
-    );
-    if (row.failed) {
-      blockDetails[index].failure += 1;
-      totalFailure += 1;
-    } else {
-      blockDetails[index].success += 1;
-      totalSuccess += 1;
-    }
-  });
-
-  blockDetails.forEach((detail) => {
-    const total = detail.success + detail.failure;
-    detail.rate = total > 0 ? detail.success / total : -1;
-  });
-  const total = totalSuccess + totalFailure;
-
-  return {
-    blockDetails,
-    successRate: total > 0 ? (totalSuccess / total) * 100 : 100,
-    totalSuccess,
-    totalFailure,
-  };
-};
-
-const DONUT_COLORS = ['#2563eb', '#22c55e', '#f97316', '#8b5cf6', '#06b6d4', '#ec4899', '#14b8a6', '#eab308'];
-
-const getDonutColor = (index: number) => DONUT_COLORS[index % DONUT_COLORS.length];
-
-type TrendPoint = {
-  key: string;
-  label: string;
-  requests: number;
-  failures: number;
-  tokens: number;
-  cost: number;
-};
-
-type TokenDistributionPoint = {
-  key: string;
-  label: string;
-  requests: number;
-  totalTokens: number;
-  inputTokens: number;
-  outputTokens: number;
-  reasoningTokens: number;
-  cachedTokens: number;
-  totalCost: number;
-};
-
-type RankingRowAccumulator = {
-  id: string;
-  group: 'apiKey' | 'model';
-  model: string;
-  apiKeyHash: string;
-  apiKeyMasked: string;
-  account: string;
-  accountMasked: string;
-  authLabels: Set<string>;
-  authIndices: Set<string>;
-  channels: Set<string>;
-  providers: Set<string>;
-  totalCalls: number;
-  successCalls: number;
-  failureCalls: number;
-  inputTokens: number;
-  outputTokens: number;
-  cachedTokens: number;
-  totalTokens: number;
-  totalCost: number;
-  latencySum: number;
-  latencyCount: number;
-  lastSeenAt: number;
-};
-
-type UsageTrendAnalytics = {
-  apiKeyOptions: Array<{ value: string; label: string }>;
-  trendPoints: TrendPoint[];
-  tokenDistributionPoints: TokenDistributionPoint[];
-  modelRows: MonitoringAccountRow[];
-  apiKeyRows: MonitoringAccountRow[];
-  scopedTotals: Record<RankingMetric, number>;
-};
-
-type AccountHealthTone = 'good' | 'warn' | 'bad';
-
-type AccountStatusBlockDetail = {
-  success: number;
-  failure: number;
-  rate: number;
-  startTime: number;
-  endTime: number;
-};
-
-type AccountStatusRange = {
-  startTime: number;
-  endTime: number;
-};
-
-type AccountStatusData = {
-  blockDetails: AccountStatusBlockDetail[];
-  successRate: number;
-  totalSuccess: number;
-  totalFailure: number;
-};
 
 type PriceTierDraft = {
 	contextSize: string;
@@ -653,255 +276,6 @@ type MonitoringSettingsDraft = {
 	webdavPassword: string;
 	modelPriceSyncEnabled: boolean;
 	modelPriceSyncIntervalMinutes: string;
-};
-
-type RealtimeLogRow = MonitoringEventRow & {
-  requestCount: number;
-  successRate: number;
-  streamKey: string;
-  diagnosticText: string;
-  errorCategoryKey: string;
-  errorSummary: string;
-  recentPattern: boolean[];
-  recentSuccessCount: number;
-  recentFailureCount: number;
-};
-
-type RealtimeCostTooltipPosition = {
-  top: number;
-  left: number;
-  arrowTop: number;
-  placement: 'left' | 'right';
-};
-
-const REALTIME_COST_TOOLTIP_WIDTH = 336;
-const REALTIME_COST_TOOLTIP_MARGIN = 12;
-
-const formatCostTierLabel = (value: string) => value
-  .trim()
-  .replace(/[-_]+/g, ' ')
-  .replace(/\b\w/g, (character) => character.toUpperCase());
-
-const calculateMillionTokenRate = (cost: number, tokens: number): number | null => (
-  tokens > 0 ? (cost / tokens) * 1_000_000 : null
-);
-
-const formatMillionTokenRate = (rate: number | null) => rate === null
-  ? '--'
-  : `$${rate.toLocaleString(undefined, { minimumFractionDigits: 4, maximumFractionDigits: 4 })} / 1M Token`;
-
-function RealtimeCostCell({ row, hasPrices, t }: { row: RealtimeLogRow; hasPrices: boolean; t: TFunction }) {
-  const cellRef = useRef<HTMLSpanElement>(null);
-  const tooltipId = useId();
-  const [tooltipPosition, setTooltipPosition] = useState<RealtimeCostTooltipPosition | null>(null);
-  const breakdown = row.costBreakdown;
-
-  const showTooltip = useCallback((element: HTMLElement | null) => {
-    if (!element || typeof window === 'undefined') return;
-    const rect = element.getBoundingClientRect();
-    const detailRowCount = breakdown
-      ? 6 + [breakdown.cacheReadTokens, breakdown.cacheWriteTokens, breakdown.reasoningTokens].filter((tokens) => tokens > 0).length
-      : 1;
-    const estimatedHeight = Math.min(420, 70 + detailRowCount * 31);
-    const placement = rect.left >= REALTIME_COST_TOOLTIP_WIDTH + REALTIME_COST_TOOLTIP_MARGIN * 2 ? 'left' : 'right';
-    const unclampedLeft = placement === 'left'
-      ? rect.left - REALTIME_COST_TOOLTIP_WIDTH - REALTIME_COST_TOOLTIP_MARGIN
-      : rect.right + REALTIME_COST_TOOLTIP_MARGIN;
-    const left = Math.min(
-      Math.max(REALTIME_COST_TOOLTIP_MARGIN, unclampedLeft),
-      Math.max(REALTIME_COST_TOOLTIP_MARGIN, window.innerWidth - REALTIME_COST_TOOLTIP_WIDTH - REALTIME_COST_TOOLTIP_MARGIN)
-    );
-    const centerY = rect.top + rect.height / 2;
-    const top = Math.min(
-      Math.max(REALTIME_COST_TOOLTIP_MARGIN, centerY - estimatedHeight / 2),
-      Math.max(REALTIME_COST_TOOLTIP_MARGIN, window.innerHeight - estimatedHeight - REALTIME_COST_TOOLTIP_MARGIN)
-    );
-    setTooltipPosition({
-      top,
-      left,
-      placement,
-      arrowTop: Math.min(Math.max(22, centerY - top), estimatedHeight - 22),
-    });
-  }, [breakdown]);
-
-  const hideTooltip = useCallback(() => setTooltipPosition(null), []);
-
-  if (!hasPrices && !breakdown) return <span>--</span>;
-
-  const conditionalCosts = breakdown ? [
-    { key: 'cache-read', tokens: breakdown.cacheReadTokens, label: t('monitoring.cost_detail_cache_read'), cost: breakdown.cacheReadCost },
-    { key: 'cache-write', tokens: breakdown.cacheWriteTokens, label: t('monitoring.cost_detail_cache_write'), cost: breakdown.cacheWriteCost },
-    { key: 'reasoning', tokens: breakdown.reasoningTokens, label: t('monitoring.cost_detail_reasoning'), cost: breakdown.reasoningCost },
-  ].filter((item) => item.tokens > 0 || item.cost > 0) : [];
-  const actualTier = breakdown?.serviceTier || row.serviceTier;
-  const actualTierLabel = actualTier
-    ? formatCostTierLabel(actualTier)
-    : t('monitoring.cost_detail_standard');
-  const billingMode = breakdown?.serviceTier
-    ? t('monitoring.cost_detail_service_tier_mode')
-    : breakdown && breakdown.contextTierSize > 0
-      ? t('monitoring.cost_detail_context_mode', { size: formatCompactNumber(breakdown.contextTierSize) })
-      : t('monitoring.cost_detail_standard');
-
-  return (
-    <span
-      ref={cellRef}
-      className={styles.realtimeCostCell}
-      onMouseEnter={() => showTooltip(cellRef.current)}
-      onMouseLeave={hideTooltip}
-    >
-      <span className={styles.realtimeCostValue}>{formatUsdPrecise(row.totalCost)}</span>
-      <button
-        type="button"
-        className={styles.realtimeCostInfoButton}
-        aria-label={t('monitoring.cost_detail_open')}
-        aria-describedby={tooltipPosition ? tooltipId : undefined}
-        onFocus={(event) => showTooltip(event.currentTarget)}
-        onBlur={hideTooltip}
-      >
-        <IconInfo size={16} />
-      </button>
-      {tooltipPosition && typeof document !== 'undefined' ? createPortal(
-        <div
-          id={tooltipId}
-          role="tooltip"
-          className={styles.realtimeCostTooltip}
-          data-placement={tooltipPosition.placement}
-          style={{
-            top: tooltipPosition.top,
-            left: tooltipPosition.left,
-            '--realtime-cost-arrow-top': `${tooltipPosition.arrowTop}px`,
-          } as CSSProperties}
-        >
-          <strong className={styles.realtimeCostTooltipTitle}>{t('monitoring.cost_detail_title')}</strong>
-          {breakdown ? (
-            <div className={styles.realtimeCostTooltipRows}>
-              <div><span>{t('monitoring.cost_detail_input')}</span><strong>{formatUsdPrecise(breakdown.inputCost)}</strong></div>
-              <div><span>{t('monitoring.cost_detail_output')}</span><strong>{formatUsdPrecise(breakdown.outputCost)}</strong></div>
-              {conditionalCosts.map((item) => (
-                <div key={item.key}><span>{item.label}</span><strong>{formatUsdPrecise(item.cost)}</strong></div>
-              ))}
-              <div className={styles.realtimeCostTooltipDivider} aria-hidden="true" />
-              <div><span>{t('monitoring.cost_detail_input_rate')}</span><strong className={styles.realtimeCostRateInput}>{formatMillionTokenRate(calculateMillionTokenRate(breakdown.inputCost, breakdown.inputTokens))}</strong></div>
-              <div><span>{t('monitoring.cost_detail_output_rate')}</span><strong className={styles.realtimeCostRateOutput}>{formatMillionTokenRate(calculateMillionTokenRate(breakdown.outputCost, breakdown.outputTokens))}</strong></div>
-              <div><span>{t('monitoring.cost_detail_actual_tier')}</span><strong>{actualTierLabel}</strong></div>
-              <div><span>{t('monitoring.cost_detail_billing_mode')}</span><strong>{billingMode}</strong></div>
-            </div>
-          ) : (
-            <p className={styles.realtimeCostTooltipEmpty}>{t('monitoring.cost_detail_unavailable')}</p>
-          )}
-        </div>,
-        document.body
-      ) : null}
-    </span>
-  );
-}
-
-type MonitoringSummaryAccumulator = {
-  totalCalls: number;
-  failureCalls: number;
-  inputTokens: number;
-  outputTokens: number;
-  reasoningTokens: number;
-  cachedTokens: number;
-  totalTokens: number;
-  totalCost: number;
-  latencySum: number;
-  latencyCount: number;
-  recentCalls: number;
-  recentTokens: number;
-  zeroTokenCalls: number;
-  taskMap: Map<string, boolean>;
-  activeDays: Set<string>;
-  zeroTokenModels: Set<string>;
-};
-
-const createMonitoringSummaryAccumulator = (): MonitoringSummaryAccumulator => ({
-  totalCalls: 0,
-  failureCalls: 0,
-  inputTokens: 0,
-  outputTokens: 0,
-  reasoningTokens: 0,
-  cachedTokens: 0,
-  totalTokens: 0,
-  totalCost: 0,
-  latencySum: 0,
-  latencyCount: 0,
-  recentCalls: 0,
-  recentTokens: 0,
-  zeroTokenCalls: 0,
-  taskMap: new Map(),
-  activeDays: new Set(),
-  zeroTokenModels: new Set(),
-});
-
-const addMonitoringSummaryRow = (
-  accumulator: MonitoringSummaryAccumulator,
-  row: MonitoringEventRow,
-  windowStartMs: number,
-  nowMs: number
-) => {
-  accumulator.totalCalls += 1;
-  if (row.failed) accumulator.failureCalls += 1;
-  accumulator.inputTokens += row.inputTokens;
-  accumulator.outputTokens += row.outputTokens;
-  accumulator.reasoningTokens += row.reasoningTokens;
-  accumulator.cachedTokens += row.cachedTokens;
-  accumulator.totalTokens += row.totalTokens;
-  accumulator.totalCost += row.totalCost;
-  accumulator.activeDays.add(row.dayKey);
-
-  if (row.latencyMs !== null) {
-    accumulator.latencySum += row.latencyMs;
-    accumulator.latencyCount += 1;
-  }
-
-  accumulator.taskMap.set(row.taskKey, (accumulator.taskMap.get(row.taskKey) ?? false) || row.failed);
-
-  if (row.totalTokens === 0) {
-    accumulator.zeroTokenCalls += 1;
-    accumulator.zeroTokenModels.add(row.model);
-  }
-
-  if (row.timestampMs >= windowStartMs && row.timestampMs <= nowMs) {
-    accumulator.recentCalls += 1;
-    accumulator.recentTokens += row.totalTokens;
-  }
-};
-
-const finalizeMonitoringSummary = (accumulator: MonitoringSummaryAccumulator): MonitoringSummary => {
-  const successCalls = Math.max(accumulator.totalCalls - accumulator.failureCalls, 0);
-  let approxTaskFailures = 0;
-  accumulator.taskMap.forEach((failed) => {
-    if (failed) approxTaskFailures += 1;
-  });
-  const activeDayCount = Math.max(accumulator.activeDays.size, 1);
-
-  return {
-    totalCalls: accumulator.totalCalls,
-    successCalls,
-    failureCalls: accumulator.failureCalls,
-    successRate: accumulator.totalCalls > 0 ? successCalls / accumulator.totalCalls : 1,
-    inputTokens: accumulator.inputTokens,
-    outputTokens: accumulator.outputTokens,
-    reasoningTokens: accumulator.reasoningTokens,
-    cachedTokens: accumulator.cachedTokens,
-    totalTokens: accumulator.totalTokens,
-    totalCost: accumulator.totalCost,
-    averageLatencyMs: accumulator.latencyCount > 0 ? accumulator.latencySum / accumulator.latencyCount : null,
-    rpm30m: accumulator.recentCalls / 30,
-    tpm30m: accumulator.recentTokens / 30,
-    avgDailyRequests: accumulator.totalCalls / activeDayCount,
-    avgDailyTokens: accumulator.totalTokens / activeDayCount,
-    approxTasks: accumulator.taskMap.size,
-    approxTaskFailures,
-    approxTaskSuccessRate:
-      accumulator.taskMap.size > 0
-        ? Math.max(accumulator.taskMap.size - approxTaskFailures, 0) / accumulator.taskMap.size
-        : 1,
-    zeroTokenCalls: accumulator.zeroTokenCalls,
-    zeroTokenModels: Array.from(accumulator.zeroTokenModels).sort(),
-  };
 };
 
 type UsageImportResult = {
@@ -977,66 +351,6 @@ const SuccessFailureValue = ({ success, failure }: { success: number; failure: n
     <span className={styles.goodText}>{formatCompactNumber(success)}</span>
     <span className={styles.badText}>({formatCompactNumber(failure)})</span>
   </span>
-);
-
-const getRankingMetricValue = (row: MonitoringAccountRow, metric: RankingMetric) => {
-  if (metric === 'cost') return row.totalCost;
-  if (metric === 'tokens') return row.totalTokens;
-  return row.totalCalls;
-};
-
-const getAccountSortValue = (row: MonitoringAccountRow, metric: AccountSortMetric) => {
-  if (metric === 'recent') return row.lastSeenAt;
-  return getRankingMetricValue(row, metric);
-};
-
-const getRankingMetricLabel = (metric: RankingMetric, t: TFunction) => {
-  if (metric === 'cost') return t('monitoring.ranking_metric_cost');
-  if (metric === 'tokens') return t('monitoring.ranking_metric_tokens');
-  return t('monitoring.ranking_metric_requests');
-};
-
-const getRankingSummaryLabel = (metric: RankingMetric, t: TFunction) => {
-  if (metric === 'cost') return t('monitoring.ranking_summary_cost');
-  if (metric === 'tokens') return t('monitoring.ranking_summary_tokens');
-  return t('monitoring.ranking_summary_calls');
-};
-
-const formatRankingMetricValue = (value: number, metric: RankingMetric, hasPrices: boolean) => {
-  if (metric === 'cost') return hasPrices ? formatUsd(value) : '--';
-  return formatCompactNumber(value);
-};
-
-const getAccountHealthTone = (row: MonitoringAccountRow): AccountHealthTone => {
-  if (row.successRate >= 0.95) return 'good';
-  if (row.successRate >= 0.85) return 'warn';
-  return 'bad';
-};
-
-const RankingMetricSwitch = ({
-  value,
-  onChange,
-  disabledCost,
-  t,
-}: {
-  value: RankingMetric;
-  onChange: (value: RankingMetric) => void;
-  disabledCost: boolean;
-  t: TFunction;
-}) => (
-  <div className={styles.rankingMetricSwitch} role="group" aria-label={t('monitoring.ranking_metric_aria')}>
-    {RANKING_METRIC_OPTIONS.map((option) => (
-      <button
-        key={option.value}
-        type="button"
-        className={`${styles.rankingMetricButton} ${value === option.value ? styles.rankingMetricButtonActive : ''}`}
-        onClick={() => onChange(option.value)}
-        disabled={option.value === 'cost' && disabledCost}
-      >
-        {t(option.labelKey)}
-      </button>
-    ))}
-  </div>
 );
 
 const buildAccountCardFileName = (row: MonitoringAccountRow, quotaEntries: AccountQuotaEntry[] = []) => {
@@ -1193,637 +507,6 @@ type AccountSummaryMetric = {
   valueClassName?: string;
 };
 
-const formatPercent = (value: number) => `${(value * 100).toFixed(1)}%`;
-
-const getProgressWidth = (value: number) => {
-  if (value <= 0) return '0%';
-  return `${Math.max(value * 100, 1.5)}%`;
-};
-
-const getChartAxisLabels = <T extends { key: string; label: string }>(points: T[]) => {
-  if (points.length <= 10) {
-    return points.map((point, index) => ({ key: point.key, label: point.label, index }));
-  }
-
-  const step = Math.ceil((points.length - 1) / 8);
-  const labels = points
-    .map((point, index) => ({ key: point.key, label: point.label, index }))
-    .filter((_, index) => index % step === 0);
-  const last = points[points.length - 1];
-  if (!labels.some((item) => item.key === last.key)) {
-    labels.push({ key: last.key, label: last.label, index: points.length - 1 });
-  }
-  return labels;
-};
-
-const buildUsageTrendRangeLabel = (range: MonitoringTimeRange, t: TFunction) => {
-  if (range === 'all') return t('monitoring.all_retained_logs');
-
-  const nowMs = Date.now();
-  return `${formatShortDateTime(getRangeStartMs(range, nowMs))} - ${formatShortDateTime(nowMs)}`;
-};
-
-const getEmptyTrendPoint = (key: string, label: string): TrendPoint => ({
-  key,
-  label,
-  requests: 0,
-  failures: 0,
-  tokens: 0,
-  cost: 0,
-});
-
-const buildFilledTrendBuckets = (range: MonitoringTimeRange, nowMs: number) => {
-  if (range === 'all') return [];
-
-  const startMs = getRangeStartMs(range, nowMs);
-  const buckets: TrendPoint[] = [];
-  const cursor = new Date(startMs);
-
-  if (range === 'today') {
-    const now = new Date(nowMs);
-    cursor.setMinutes(0, 0, 0);
-    while (cursor.getTime() <= now.getTime()) {
-      const dayKey = buildLocalDayKey(cursor.getTime());
-      const label = buildHourLabel(cursor.getTime());
-      buckets.push(getEmptyTrendPoint(`${dayKey} ${label}`, label));
-      cursor.setHours(cursor.getHours() + 1);
-    }
-    return buckets;
-  }
-
-  cursor.setHours(0, 0, 0, 0);
-  const end = new Date(nowMs);
-  end.setHours(0, 0, 0, 0);
-  while (cursor.getTime() <= end.getTime()) {
-    const key = buildLocalDayKey(cursor.getTime());
-    buckets.push(getEmptyTrendPoint(key, buildDayLabel(key)));
-    cursor.setDate(cursor.getDate() + 1);
-  }
-  return buckets;
-};
-
-const buildTimeBucketMeta = (range: MonitoringTimeRange) => {
-  const useHourly = range === 'today';
-  return {
-    useHourly,
-    getKey: (row: MonitoringEventRow) => (useHourly ? `${row.dayKey} ${row.hourLabel}` : row.dayKey),
-    getLabel: (row: MonitoringEventRow) => (useHourly ? row.hourLabel : buildDayLabel(row.dayKey)),
-  };
-};
-
-const createRankingRowAccumulator = (
-  row: MonitoringEventRow,
-  group: 'apiKey' | 'model'
-): RankingRowAccumulator => {
-  if (group === 'apiKey') {
-    return {
-      id: row.clientApiKey.id,
-      group,
-      model: '-',
-      apiKeyHash: row.clientApiKey.hash,
-      apiKeyMasked: row.clientApiKey.masked,
-      account: row.clientApiKey.masked,
-      accountMasked: row.clientApiKey.masked,
-      authLabels: new Set(),
-      authIndices: new Set(),
-      channels: new Set(),
-      providers: new Set(),
-      totalCalls: 0,
-      successCalls: 0,
-      failureCalls: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      cachedTokens: 0,
-      totalTokens: 0,
-      totalCost: 0,
-      latencySum: 0,
-      latencyCount: 0,
-      lastSeenAt: 0,
-    };
-  }
-
-  return {
-    id: `model:${row.model}`,
-    group,
-    model: row.model,
-    apiKeyHash: '-',
-    apiKeyMasked: '-',
-    account: row.model,
-    accountMasked: row.model,
-    authLabels: new Set(),
-    authIndices: new Set(),
-    channels: new Set(),
-    providers: new Set(),
-    totalCalls: 0,
-    successCalls: 0,
-    failureCalls: 0,
-    inputTokens: 0,
-    outputTokens: 0,
-    cachedTokens: 0,
-    totalTokens: 0,
-    totalCost: 0,
-    latencySum: 0,
-    latencyCount: 0,
-    lastSeenAt: 0,
-  };
-};
-
-const addRankingRow = (accumulator: RankingRowAccumulator, row: MonitoringEventRow) => {
-  accumulator.authLabels.add(row.authLabel);
-  accumulator.authIndices.add(row.authIndexMasked);
-  accumulator.channels.add(row.channel);
-  accumulator.providers.add(row.provider);
-  accumulator.totalCalls += 1;
-  accumulator.successCalls += row.failed ? 0 : 1;
-  accumulator.failureCalls += row.failed ? 1 : 0;
-  accumulator.inputTokens += row.inputTokens;
-  accumulator.outputTokens += row.outputTokens;
-  accumulator.cachedTokens += row.cachedTokens;
-  accumulator.totalTokens += row.totalTokens;
-  accumulator.totalCost += row.totalCost;
-  accumulator.lastSeenAt = Math.max(accumulator.lastSeenAt, row.timestampMs);
-
-  if (row.latencyMs !== null) {
-    accumulator.latencySum += row.latencyMs;
-    accumulator.latencyCount += 1;
-  }
-};
-
-const finalizeRankingRows = (grouped: Map<string, RankingRowAccumulator>): MonitoringAccountRow[] =>
-  Array.from(grouped.values()).map((item) => ({
-    id: item.id,
-    group: item.group,
-    model: item.model,
-    apiKeyHash: item.apiKeyHash,
-    apiKeyMasked: item.apiKeyMasked,
-    account: item.account,
-    accountMasked: item.accountMasked,
-    authLabels: Array.from(item.authLabels).sort(),
-    authIndices: Array.from(item.authIndices).sort(),
-    channels: Array.from(item.channels).sort(),
-    providers: Array.from(item.providers).sort(),
-    totalCalls: item.totalCalls,
-    successCalls: item.successCalls,
-    failureCalls: item.failureCalls,
-    successRate: item.totalCalls > 0 ? item.successCalls / item.totalCalls : 1,
-    inputTokens: item.inputTokens,
-    outputTokens: item.outputTokens,
-    cachedTokens: item.cachedTokens,
-    totalTokens: item.totalTokens,
-    totalCost: item.totalCost,
-    averageLatencyMs: item.latencyCount > 0 ? item.latencySum / item.latencyCount : null,
-    lastSeenAt: item.lastSeenAt,
-    recentPattern: [],
-    models: [],
-  }));
-
-const buildUsageTrendAnalytics = (
-  rows: MonitoringEventRow[],
-  range: MonitoringTimeRange,
-  apiKeyFilter: string,
-  allApiKeyLabel: string
-): UsageTrendAnalytics => {
-  const nowMs = Date.now();
-  const prefilled = buildFilledTrendBuckets(range, nowMs);
-  const trendGrouped = new Map<string, TrendPoint>(prefilled.map((point) => [point.key, point]));
-  const tokenGrouped = new Map<string, TokenDistributionPoint>();
-  const modelGrouped = new Map<string, RankingRowAccumulator>();
-  const apiKeyGrouped = new Map<string, RankingRowAccumulator>();
-  const apiKeyLabels = new Map<string, string>();
-  const { getKey, getLabel } = buildTimeBucketMeta(range);
-  const scopedTotals: Record<RankingMetric, number> = {
-    requests: 0,
-    tokens: 0,
-    cost: 0,
-  };
-
-  rows.forEach((row) => {
-    const apiKeyHash = row.clientApiKey.hash;
-    if (apiKeyHash && apiKeyHash !== '-') {
-      apiKeyLabels.set(apiKeyHash, row.clientApiKey.masked);
-    }
-
-    const apiKeyAccumulator = apiKeyGrouped.get(row.clientApiKey.id) ?? createRankingRowAccumulator(row, 'apiKey');
-    addRankingRow(apiKeyAccumulator, row);
-    apiKeyGrouped.set(apiKeyAccumulator.id, apiKeyAccumulator);
-
-    if (apiKeyFilter !== 'all' && apiKeyHash !== apiKeyFilter) {
-      return;
-    }
-
-    const key = getKey(row);
-    const label = getLabel(row);
-    const trendPoint = trendGrouped.get(key) ?? {
-      key,
-      label,
-      requests: 0,
-      failures: 0,
-      tokens: 0,
-      cost: 0,
-    };
-    trendPoint.requests += 1;
-    trendPoint.failures += row.failed ? 1 : 0;
-    trendPoint.tokens += row.totalTokens;
-    trendPoint.cost += row.totalCost;
-    trendGrouped.set(key, trendPoint);
-
-    const tokenPoint = tokenGrouped.get(key) ?? {
-      key,
-      label,
-      requests: 0,
-      totalTokens: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      reasoningTokens: 0,
-      cachedTokens: 0,
-      totalCost: 0,
-    };
-    tokenPoint.requests += 1;
-    tokenPoint.totalTokens += row.totalTokens;
-    tokenPoint.inputTokens += row.inputTokens;
-    tokenPoint.outputTokens += row.outputTokens;
-    tokenPoint.reasoningTokens += row.reasoningTokens;
-    tokenPoint.cachedTokens += row.cachedTokens;
-    tokenPoint.totalCost += row.totalCost;
-    tokenGrouped.set(key, tokenPoint);
-
-    const modelAccumulator = modelGrouped.get(row.model) ?? createRankingRowAccumulator(row, 'model');
-    addRankingRow(modelAccumulator, row);
-    modelGrouped.set(row.model, modelAccumulator);
-
-    scopedTotals.requests += 1;
-    scopedTotals.tokens += row.totalTokens;
-    scopedTotals.cost += row.totalCost;
-  });
-
-  const apiKeyOptions = [
-    { value: 'all', label: allApiKeyLabel },
-    ...Array.from(apiKeyLabels.entries())
-      .sort((left, right) => left[1].localeCompare(right[1]))
-      .map(([value, label]) => ({ value, label })),
-  ];
-
-  return {
-    apiKeyOptions,
-    trendPoints: Array.from(trendGrouped.values()).sort((left, right) => left.key.localeCompare(right.key)).slice(-24),
-    tokenDistributionPoints: Array.from(tokenGrouped.values()).sort((left, right) => left.key.localeCompare(right.key)).slice(-24),
-    modelRows: finalizeRankingRows(modelGrouped),
-    apiKeyRows: finalizeRankingRows(apiKeyGrouped),
-    scopedTotals,
-  };
-};
-
-const calculateAggregateCost = (
-	 item: Pick<UsageAggregateBucket, 'model' | 'inputTokens' | 'outputTokens' | 'cacheTokens' | 'estimatedCost'>,
-	 modelPrices: Record<string, ModelPrice>
-) => Number.isFinite(Number(item.estimatedCost)) && Number(item.estimatedCost) >= 0
-	 ? Number(item.estimatedCost)
-	 : calculateCost({
-	 __modelName: item.model || '',
-  tokens: {
-    input_tokens: item.inputTokens,
-    output_tokens: item.outputTokens,
-    cached_tokens: item.cacheTokens,
-    cache_tokens: item.cacheTokens,
-  },
-}, modelPrices);
-
-const createAggregateRankingRow = (
-  item: UsageAggregateBucket,
-  group: 'apiKey' | 'model',
-  modelPrices: Record<string, ModelPrice>,
-  apiKeyLabels: Map<string, string>
-): MonitoringAccountRow => {
-  const model = item.model || '-';
-  const apiKeyHash = item.apiKeyHash || '-';
-  const apiKeyMasked = apiKeyLabels.get(apiKeyHash) || maskSensitiveText(apiKeyHash);
-  const totalCost = calculateAggregateCost(item, modelPrices);
-  return {
-    id: group === 'model' ? `model:${model}` : `clientApiKey:${apiKeyHash}`,
-    group,
-    model: group === 'model' ? model : '-',
-    apiKeyHash: group === 'apiKey' ? apiKeyHash : '-',
-    apiKeyMasked: group === 'apiKey' ? apiKeyMasked : '-',
-    account: group === 'model' ? model : apiKeyMasked,
-    accountMasked: group === 'model' ? model : apiKeyMasked,
-    authLabels: [],
-    authIndices: [],
-    channels: [],
-    providers: item.provider ? [item.provider] : [],
-    totalCalls: item.totalRequests,
-    successCalls: item.successCount,
-    failureCalls: item.failureCount,
-    successRate: item.totalRequests > 0 ? item.successCount / item.totalRequests : 1,
-    inputTokens: item.inputTokens,
-    outputTokens: item.outputTokens,
-    cachedTokens: item.cacheTokens,
-    totalTokens: item.totalTokens,
-    totalCost,
-    averageLatencyMs: item.avgLatencyMs ?? null,
-    lastSeenAt: item.bucketStartMs,
-    recentPattern: [],
-    models: [],
-  };
-};
-
-const buildServerUsageTrendAnalytics = (
-  aggregates: ReturnType<typeof useUsageAggregates>['data'],
-  range: MonitoringTimeRange,
-  modelPrices: Record<string, ModelPrice>,
-  apiKeyOptions: Array<{ value: string; label: string }>,
-  apiKeyFilter: string,
-  unattributedApiKeyLabel: string
-): UsageTrendAnalytics | null => {
-  if (!aggregates) return null;
-  const nowMs = Date.now();
-  const prefilled = buildFilledTrendBuckets(range, nowMs);
-  const trendGrouped = new Map<string, TrendPoint>(prefilled.map((point) => [point.key, point]));
-  const tokenGrouped = new Map<string, TokenDistributionPoint>();
-  const apiKeyLabels = new Map(apiKeyOptions.map((option) => [option.value, option.label]));
-  apiKeyLabels.set('-', unattributedApiKeyLabel);
-  aggregates.apiKeys.forEach((item) => {
-    const apiKeyHash = item.apiKeyHash?.trim();
-    if (apiKeyHash && !apiKeyLabels.has(apiKeyHash)) {
-      apiKeyLabels.set(apiKeyHash, maskSensitiveText(apiKeyHash));
-    }
-  });
-  const resolvedApiKeyOptions = [
-    apiKeyOptions.find((option) => option.value === 'all') ?? { value: 'all', label: 'All' },
-    ...Array.from(apiKeyLabels.entries())
-      .filter(([value]) => value !== 'all' && value !== '-')
-      .sort((left, right) => left[1].localeCompare(right[1]))
-      .map(([value, label]) => ({ value, label })),
-  ];
-  const modelRowMap = new Map<string, MonitoringAccountRow>();
-  aggregates.models
-    .filter((item) => apiKeyFilter === 'all' || item.apiKeyHash === apiKeyFilter)
-    .forEach((item) => {
-      const row = createAggregateRankingRow(item, 'model', modelPrices, apiKeyLabels);
-      const current = modelRowMap.get(row.model);
-      if (!current) {
-        modelRowMap.set(row.model, row);
-        return;
-      }
-      const previousCalls = current.totalCalls;
-      current.totalCalls += row.totalCalls;
-      current.successCalls += row.successCalls;
-      current.failureCalls += row.failureCalls;
-      current.inputTokens += row.inputTokens;
-      current.outputTokens += row.outputTokens;
-      current.cachedTokens += row.cachedTokens;
-      current.totalTokens += row.totalTokens;
-      current.totalCost += row.totalCost;
-      current.lastSeenAt = Math.max(current.lastSeenAt, row.lastSeenAt);
-      if (row.averageLatencyMs !== null) {
-        const weightedCurrent = (current.averageLatencyMs ?? 0) * previousCalls;
-        current.averageLatencyMs = (weightedCurrent + row.averageLatencyMs * row.totalCalls) / Math.max(current.totalCalls, 1);
-      }
-      current.successRate = current.totalCalls > 0 ? current.successCalls / current.totalCalls : 1;
-    });
-  const modelRows = Array.from(modelRowMap.values());
-  const apiKeyRowMap = new Map<string, MonitoringAccountRow>();
-  aggregates.apiKeys.forEach((item) => {
-    const row = createAggregateRankingRow(item, 'apiKey', modelPrices, apiKeyLabels);
-    const current = apiKeyRowMap.get(row.apiKeyHash);
-    if (!current) {
-      apiKeyRowMap.set(row.apiKeyHash, row);
-      return;
-    }
-    current.totalCalls += row.totalCalls;
-    current.successCalls += row.successCalls;
-    current.failureCalls += row.failureCalls;
-    current.inputTokens += row.inputTokens;
-    current.outputTokens += row.outputTokens;
-    current.cachedTokens += row.cachedTokens;
-    current.totalTokens += row.totalTokens;
-    current.totalCost += row.totalCost;
-    current.lastSeenAt = Math.max(current.lastSeenAt, row.lastSeenAt);
-    current.successRate = current.totalCalls > 0 ? current.successCalls / current.totalCalls : 1;
-  });
-  const apiKeyRows = Array.from(apiKeyRowMap.values());
-
-  aggregates.trend.forEach((item) => {
-    const timestampMs = Number(item.bucketStartMs) || Date.parse(item.bucketStart);
-    const dayKey = buildLocalDayKey(timestampMs);
-    const useHourly = range === 'today';
-    const key = useHourly ? `${dayKey} ${buildHourLabel(timestampMs)}` : dayKey;
-    const label = useHourly ? buildHourLabel(timestampMs) : buildDayLabel(dayKey);
-    const cost = calculateAggregateCost(item, modelPrices);
-    const trendPoint = trendGrouped.get(key) ?? getEmptyTrendPoint(key, label);
-    trendPoint.requests += item.totalRequests;
-    trendPoint.failures += item.failureCount;
-    trendPoint.tokens += item.totalTokens;
-    trendPoint.cost += cost;
-    trendGrouped.set(key, trendPoint);
-
-    const tokenPoint = tokenGrouped.get(key) ?? {
-      key,
-      label,
-      requests: 0,
-      totalTokens: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      reasoningTokens: 0,
-      cachedTokens: 0,
-      totalCost: 0,
-    };
-    tokenPoint.requests += item.totalRequests;
-    tokenPoint.totalTokens += item.totalTokens;
-    tokenPoint.inputTokens += item.inputTokens;
-    tokenPoint.outputTokens += item.outputTokens;
-    tokenPoint.reasoningTokens += item.reasoningTokens;
-    tokenPoint.cachedTokens += item.cacheTokens;
-    tokenPoint.totalCost += cost;
-    tokenGrouped.set(key, tokenPoint);
-  });
-
-  const scopedTotals = modelRows.reduce<Record<RankingMetric, number>>((totals, row) => ({
-    requests: totals.requests + row.totalCalls,
-    tokens: totals.tokens + row.totalTokens,
-    cost: totals.cost + row.totalCost,
-  }), { requests: 0, tokens: 0, cost: 0 });
-
-  return {
-    apiKeyOptions: resolvedApiKeyOptions,
-    trendPoints: Array.from(trendGrouped.values()).sort((left, right) => left.key.localeCompare(right.key)).slice(-24),
-    tokenDistributionPoints: Array.from(tokenGrouped.values()).sort((left, right) => left.key.localeCompare(right.key)).slice(-24),
-    modelRows,
-    apiKeyRows,
-    scopedTotals,
-  };
-};
-
-const buildAggregateSummary = (
-  buckets: UsageAggregateBucket[],
-  modelPrices: Record<string, ModelPrice>
-): MonitoringSummary => {
-  let totalCalls = 0;
-  let successCalls = 0;
-  let failureCalls = 0;
-  let inputTokens = 0;
-  let outputTokens = 0;
-  let reasoningTokens = 0;
-  let cachedTokens = 0;
-  let totalTokens = 0;
-  let totalCost = 0;
-  let weightedLatency = 0;
-  let latencyCalls = 0;
-  buckets.forEach((bucket) => {
-    totalCalls += bucket.totalRequests;
-    successCalls += bucket.successCount;
-    failureCalls += bucket.failureCount;
-    inputTokens += bucket.inputTokens;
-    outputTokens += bucket.outputTokens;
-    reasoningTokens += bucket.reasoningTokens;
-    cachedTokens += bucket.cacheTokens;
-    totalTokens += bucket.totalTokens;
-    totalCost += calculateAggregateCost(bucket, modelPrices);
-    if (typeof bucket.avgLatencyMs === 'number' && bucket.totalRequests > 0) {
-      weightedLatency += bucket.avgLatencyMs * bucket.totalRequests;
-      latencyCalls += bucket.totalRequests;
-    }
-  });
-  return {
-    totalCalls,
-    successCalls,
-    failureCalls,
-    successRate: totalCalls > 0 ? successCalls / totalCalls : 1,
-    inputTokens,
-    outputTokens,
-    reasoningTokens,
-    cachedTokens,
-    totalTokens,
-    totalCost,
-    averageLatencyMs: latencyCalls > 0 ? weightedLatency / latencyCalls : null,
-    rpm30m: 0,
-    tpm30m: 0,
-    avgDailyRequests: 0,
-    avgDailyTokens: 0,
-    approxTasks: 0,
-    approxTaskFailures: 0,
-    approxTaskSuccessRate: 1,
-    zeroTokenCalls: 0,
-    zeroTokenModels: [],
-  };
-};
-
-const buildServerAccountRows = (
-  buckets: UsageAggregateBucket[],
-  realtimeRows: MonitoringEventRow[],
-  authFilesByAuthIndex: Map<string, AuthFileItem>,
-  modelPrices: Record<string, ModelPrice>,
-  deletedCredentialLabel: string
-): MonitoringAccountRow[] => {
-  const metadataByAuthIndex = new Map<string, MonitoringEventRow>();
-  const realtimeRowsByAuthIndex = new Map<string, MonitoringEventRow[]>();
-  realtimeRows.forEach((row) => {
-    if (row.authIndex !== '-' && !metadataByAuthIndex.has(row.authIndex)) {
-      metadataByAuthIndex.set(row.authIndex, row);
-    }
-    if (row.authIndex !== '-') {
-      const items = realtimeRowsByAuthIndex.get(row.authIndex) ?? [];
-      items.push(row);
-      realtimeRowsByAuthIndex.set(row.authIndex, items);
-    }
-  });
-  const grouped = new Map<string, MonitoringAccountRow>();
-  buckets.forEach((bucket) => {
-    const authIndex = normalizeAuthIndex(bucket.authIndex) ?? '-';
-    const metadata = metadataByAuthIndex.get(authIndex);
-    const authFile = authFilesByAuthIndex.get(authIndex);
-    const authFileLabel = authFile
-      ? [authFile.email, authFile.account, authFile.label, authFile.name]
-          .map((value) => typeof value === 'string' ? value.trim() : '')
-          .find(Boolean) || ''
-      : '';
-    const fallbackAccount = authIndex === '-' ? deletedCredentialLabel : maskSensitiveText(authIndex);
-    const account = metadata?.account || metadata?.authLabel || authFileLabel || fallbackAccount;
-    const accountMasked = metadata?.accountMasked || metadata?.authIndexMasked || maskSensitiveText(account);
-    const provider = bucket.provider || metadata?.provider || '-';
-    const channel = metadata?.channel || provider;
-    const id = `account:${account}::${channel}`;
-    const current = grouped.get(id) ?? {
-      id,
-      group: 'account',
-      model: '-',
-      apiKeyHash: '-',
-      apiKeyMasked: '-',
-      account,
-      accountMasked,
-      authLabels: [],
-      authIndices: [],
-      channels: [],
-      providers: [],
-      totalCalls: 0,
-      successCalls: 0,
-      failureCalls: 0,
-      successRate: 1,
-      inputTokens: 0,
-      outputTokens: 0,
-      cachedTokens: 0,
-      totalTokens: 0,
-      totalCost: 0,
-      averageLatencyMs: null,
-      lastSeenAt: 0,
-      recentPattern: [],
-      rows: [],
-      models: [],
-    } satisfies MonitoringAccountRow;
-    current.totalCalls += bucket.totalRequests;
-    current.successCalls += bucket.successCount;
-    current.failureCalls += bucket.failureCount;
-    current.inputTokens += bucket.inputTokens;
-    current.outputTokens += bucket.outputTokens;
-    current.cachedTokens += bucket.cacheTokens;
-    current.totalTokens += bucket.totalTokens;
-    current.totalCost += calculateAggregateCost(bucket, modelPrices);
-    current.lastSeenAt = Math.max(current.lastSeenAt, Number(bucket.lastSeenAtMs) || bucket.bucketStartMs || 0);
-    current.successRate = current.totalCalls > 0 ? current.successCalls / current.totalCalls : 1;
-    current.authLabels = Array.from(new Set([...current.authLabels, metadata?.authLabel || account]));
-    current.authIndices = Array.from(new Set([...current.authIndices, metadata?.authIndexMasked || maskSensitiveText(authIndex)]));
-    current.channels = Array.from(new Set([...current.channels, channel]));
-    current.providers = Array.from(new Set([...current.providers, provider]));
-    current.rows = Array.from(new Map([
-      ...(current.rows ?? []).map((row) => [row.id, row] as const),
-      ...(realtimeRowsByAuthIndex.get(authIndex) ?? []).map((row) => [row.id, row] as const),
-    ]).values());
-    current.recentPattern = (current.rows ?? []).slice(0, 10).reverse().map((row) => !row.failed);
-    const existingModel = current.models.find((item) => item.model === (bucket.model || '-'));
-    const modelCost = calculateAggregateCost(bucket, modelPrices);
-    if (existingModel) {
-      existingModel.totalCalls += bucket.totalRequests;
-      existingModel.successCalls += bucket.successCount;
-      existingModel.failureCalls += bucket.failureCount;
-      existingModel.inputTokens += bucket.inputTokens;
-      existingModel.outputTokens += bucket.outputTokens;
-      existingModel.cachedTokens += bucket.cacheTokens;
-      existingModel.totalTokens += bucket.totalTokens;
-      existingModel.totalCost += modelCost;
-      existingModel.lastSeenAt = Math.max(existingModel.lastSeenAt, Number(bucket.lastSeenAtMs) || 0);
-      existingModel.successRate = existingModel.totalCalls > 0 ? existingModel.successCalls / existingModel.totalCalls : 1;
-    } else {
-      current.models.push({
-        model: bucket.model || '-',
-        totalCalls: bucket.totalRequests,
-        successCalls: bucket.successCount,
-        failureCalls: bucket.failureCount,
-        successRate: bucket.totalRequests > 0 ? bucket.successCount / bucket.totalRequests : 1,
-        inputTokens: bucket.inputTokens,
-        outputTokens: bucket.outputTokens,
-        cachedTokens: bucket.cacheTokens,
-        totalTokens: bucket.totalTokens,
-        totalCost: modelCost,
-        lastSeenAt: Number(bucket.lastSeenAtMs) || 0,
-      });
-    }
-    grouped.set(id, current);
-  });
-  return Array.from(grouped.values()).map((row) => ({
-    ...row,
-    models: [...row.models].sort((left, right) => right.totalCost - left.totalCost || right.totalCalls - left.totalCalls),
-  }));
-};
-
 const roundCurrency = (value: number) => Math.round(value * 100) / 100;
 
 const formatDeltaPercent = (current: number, previous: number) => {
@@ -1864,246 +547,6 @@ const MODEL_PRICE_SYNC_RATE_FIELDS = [
   ['cacheRead', 'usage_stats.model_price_cache_read'],
   ['cacheWrite', 'usage_stats.model_price_cache_write'],
 ] as const;
-
-const buildRealtimeMetaText = (row: MonitoringEventRow) => {
-  const parts = [`${row.endpointMethod} ${row.endpointPath}`.trim()];
-  const text = parts.filter(Boolean).join(' · ');
-  return maskSensitiveText(text || '-');
-};
-
-const buildRealtimeDiagnosticText = (row: MonitoringEventRow) => {
-  const parts: string[] = [];
-  if (row.statusCode !== null && row.statusCode >= 400) {
-    parts.push(`HTTP ${row.statusCode}`);
-  }
-  if (row.errorCode) parts.push(row.errorCode);
-  if (row.retryAfter) parts.push(`Retry ${row.retryAfter}`);
-  return maskSensitiveText(parts.join(' · '));
-};
-
-const buildRealtimeStatusCodeText = (row: Pick<MonitoringEventRow, 'statusCode' | 'errorCode'>) => {
-  if (row.statusCode !== null && row.statusCode >= 400) return String(row.statusCode);
-  return row.errorCode ? maskSensitiveText(row.errorCode) : '';
-};
-
-const buildRealtimeStatusLabel = (
-  row: Pick<MonitoringEventRow, 'failed' | 'statusCode' | 'errorCode'>,
-  label: string
-) => {
-  if (!row.failed) return label;
-  const codeText = buildRealtimeStatusCodeText(row);
-  return codeText ? `${label} · ${codeText}` : label;
-};
-
-const compactRealtimeErrorMessage = (message: string, maxLength = 220) => {
-  const masked = maskSensitiveText(message.replace(/\s+/g, ' ').trim());
-  return masked.length > maxLength ? `${masked.slice(0, maxLength - 1)}...` : masked;
-};
-
-const resolveRealtimeErrorCategoryKey = (row: MonitoringEventRow) => {
-  const code = row.errorCode.toLowerCase();
-  const message = row.errorMessage.toLowerCase();
-  const status = row.statusCode;
-  const combined = `${code} ${message}`;
-
-  if (status === 401 || status === 403 || /\b(auth|unauthorized|forbidden|invalid[_ -]?key|permission)\b/.test(combined)) {
-    return 'monitoring.error_category_auth';
-  }
-  if (status === 429 || /\b(rate[_ -]?limit|too many requests|quota|insufficient_quota)\b/.test(combined)) {
-    return 'monitoring.error_category_rate_limit';
-  }
-  if (status === 400 || /\b(bad[_ -]?request|invalid[_ -]?request|validation)\b/.test(combined)) {
-    return 'monitoring.error_category_bad_request';
-  }
-  if (status === 404 || /\b(model.*not.*found|not[_ -]?found|404)\b/.test(combined)) {
-    return 'monitoring.error_category_not_found';
-  }
-  if (/\b(timeout|deadline|context canceled|connection reset|econnreset|network)\b/.test(combined)) {
-    return 'monitoring.error_category_network';
-  }
-  if (status !== null && status >= 500) {
-    return 'monitoring.error_category_upstream';
-  }
-  return row.failed ? 'monitoring.error_category_unknown' : 'monitoring.error_category_none';
-};
-
-const REALTIME_ERROR_TEXT_FALLBACKS = {
-  en: {
-    error_details: 'Error Details',
-    error_details_click_hint: 'Click to view error details',
-    error_details_modal_desc: 'Only fields directly related to the failed request are shown here.',
-    error_category: 'Error Category',
-    error_category_none: 'No Error',
-    error_category_auth: 'Auth / Permission',
-    error_category_rate_limit: 'Rate Limit / Quota',
-    error_category_bad_request: 'Bad Request',
-    error_category_not_found: 'Not Found',
-    error_category_network: 'Network / Timeout',
-    error_category_upstream: 'Upstream Error',
-    error_category_unknown: 'Unknown Error',
-    http_status: 'HTTP Status',
-    error_code: 'Error Code',
-    error_message: 'Error Message',
-    upstream_request_id: 'Upstream Request ID',
-    retry_after: 'Retry After',
-    copy_diagnostic: 'Copy Diagnostic',
-    copy_diagnostic_success: 'Diagnostic copied',
-    copy_diagnostic_failed: 'Unable to copy diagnostic',
-    request_status: 'Request Status',
-    filter_provider: 'Provider',
-    column_model: 'Model',
-  },
-  ru: {
-    error_details: 'Детали ошибки',
-    error_details_click_hint: 'Нажмите, чтобы посмотреть детали ошибки',
-    error_details_modal_desc: 'Здесь показаны только поля, напрямую связанные с ошибкой запроса.',
-    error_category: 'Категория ошибки',
-    error_category_none: 'Нет ошибки',
-    error_category_auth: 'Авторизация / права',
-    error_category_rate_limit: 'Лимит / квота',
-    error_category_bad_request: 'Некорректный запрос',
-    error_category_not_found: 'Не найдено',
-    error_category_network: 'Сеть / тайм-аут',
-    error_category_upstream: 'Ошибка upstream',
-    error_category_unknown: 'Неизвестная ошибка',
-    http_status: 'HTTP статус',
-    error_code: 'Код ошибки',
-    error_message: 'Сообщение ошибки',
-    upstream_request_id: 'Upstream request ID',
-    retry_after: 'Повторить после',
-    copy_diagnostic: 'Скопировать диагностику',
-    copy_diagnostic_success: 'Диагностика скопирована',
-    copy_diagnostic_failed: 'Не удалось скопировать диагностику',
-    request_status: 'Статус запроса',
-    filter_provider: 'Провайдер',
-    column_model: 'Модель',
-  },
-  zhCN: {
-    error_details: '错误详情',
-    error_details_click_hint: '点击查看错误详情',
-    error_details_modal_desc: '这里只显示和本次请求失败直接相关的字段。',
-    error_category: '错误类别',
-    error_category_none: '无错误',
-    error_category_auth: '鉴权 / 权限',
-    error_category_rate_limit: '限流 / 配额',
-    error_category_bad_request: '请求参数错误',
-    error_category_not_found: '资源不存在',
-    error_category_network: '网络 / 超时',
-    error_category_upstream: '上游错误',
-    error_category_unknown: '未知错误',
-    http_status: 'HTTP 状态',
-    error_code: '错误码',
-    error_message: '错误信息',
-    upstream_request_id: '上游请求 ID',
-    retry_after: '重试等待',
-    copy_diagnostic: '复制诊断',
-    copy_diagnostic_success: '诊断信息已复制',
-    copy_diagnostic_failed: '无法复制诊断信息',
-    request_status: '请求状态',
-    filter_provider: '提供商',
-    column_model: '模型',
-  },
-  zhTW: {
-    error_details: '錯誤詳情',
-    error_details_click_hint: '點擊查看錯誤詳情',
-    error_details_modal_desc: '這裡只顯示與本次請求失敗直接相關的欄位。',
-    error_category: '錯誤類別',
-    error_category_none: '無錯誤',
-    error_category_auth: '驗證 / 權限',
-    error_category_rate_limit: '限流 / 配額',
-    error_category_bad_request: '請求參數錯誤',
-    error_category_not_found: '資源不存在',
-    error_category_network: '網路 / 逾時',
-    error_category_upstream: '上游錯誤',
-    error_category_unknown: '未知錯誤',
-    http_status: 'HTTP 狀態',
-    error_code: '錯誤碼',
-    error_message: '錯誤訊息',
-    upstream_request_id: '上游請求 ID',
-    retry_after: '重試等待',
-    copy_diagnostic: '複製診斷',
-    copy_diagnostic_success: '診斷資訊已複製',
-    copy_diagnostic_failed: '無法複製診斷資訊',
-    request_status: '請求狀態',
-    filter_provider: '提供商',
-    column_model: '模型',
-  },
-} as const;
-
-type RealtimeErrorTextKey = keyof typeof REALTIME_ERROR_TEXT_FALLBACKS.en;
-
-const resolveRealtimeErrorFallbackLocale = (language?: string) => {
-  const normalized = language?.toLowerCase() ?? '';
-  if (normalized.startsWith('zh-tw') || normalized.startsWith('zh-hk') || normalized.startsWith('zh-mo')) return 'zhTW';
-  if (normalized.startsWith('zh')) return 'zhCN';
-  if (normalized.startsWith('ru')) return 'ru';
-  return 'en';
-};
-
-const translateRealtimeErrorText = (
-  key: RealtimeErrorTextKey,
-  t: ReturnType<typeof useTranslation>['t'],
-  language?: string
-) => {
-  const fallbackLocale = resolveRealtimeErrorFallbackLocale(language);
-  const fallback = REALTIME_ERROR_TEXT_FALLBACKS[fallbackLocale][key] ?? REALTIME_ERROR_TEXT_FALLBACKS.en[key];
-  return t(`monitoring.${key}`, { defaultValue: fallback });
-};
-
-const translateRealtimeErrorCategory = (
-  key: string,
-  t: ReturnType<typeof useTranslation>['t'],
-  language?: string
-) => {
-  switch (key) {
-    case 'monitoring.error_category_auth':
-      return translateRealtimeErrorText('error_category_auth', t, language);
-    case 'monitoring.error_category_rate_limit':
-      return translateRealtimeErrorText('error_category_rate_limit', t, language);
-    case 'monitoring.error_category_bad_request':
-      return translateRealtimeErrorText('error_category_bad_request', t, language);
-    case 'monitoring.error_category_not_found':
-      return translateRealtimeErrorText('error_category_not_found', t, language);
-    case 'monitoring.error_category_network':
-      return translateRealtimeErrorText('error_category_network', t, language);
-    case 'monitoring.error_category_upstream':
-      return translateRealtimeErrorText('error_category_upstream', t, language);
-    case 'monitoring.error_category_none':
-      return translateRealtimeErrorText('error_category_none', t, language);
-    case 'monitoring.error_category_unknown':
-    default:
-      return translateRealtimeErrorText('error_category_unknown', t, language);
-  }
-};
-
-const buildRealtimeErrorSummary = (row: MonitoringEventRow) => {
-  if (!row.failed) return '';
-  const parts: string[] = [];
-  if (row.errorMessage) parts.push(compactRealtimeErrorMessage(row.errorMessage));
-  if (!row.errorMessage && row.errorCode) parts.push(maskSensitiveText(row.errorCode));
-  if (row.upstreamRequestId) parts.push(`RID ${maskSensitiveText(row.upstreamRequestId)}`);
-  if (row.retryAfter) parts.push(`Retry ${maskSensitiveText(row.retryAfter)}`);
-  return parts.join(' · ');
-};
-
-const buildRealtimeDiagnosticClipboardText = (
-  row: RealtimeLogRow,
-  t: ReturnType<typeof useTranslation>['t'],
-  language?: string
-) => {
-  const fields: Array<[string, string | number | null | undefined]> = [
-    [translateRealtimeErrorText('request_status', t, language), row.failed ? t('monitoring.result_failed') : t('monitoring.result_success')],
-    [translateRealtimeErrorText('error_category', t, language), translateRealtimeErrorCategory(row.errorCategoryKey, t, language)],
-    [translateRealtimeErrorText('http_status', t, language), row.statusCode ?? '-'],
-    [translateRealtimeErrorText('error_code', t, language), row.errorCode || '-'],
-    [translateRealtimeErrorText('error_message', t, language), row.errorMessage ? compactRealtimeErrorMessage(row.errorMessage, 800) : '-'],
-    [translateRealtimeErrorText('upstream_request_id', t, language), row.upstreamRequestId || '-'],
-    [translateRealtimeErrorText('retry_after', t, language), row.retryAfter || '-'],
-    [translateRealtimeErrorText('filter_provider', t, language), row.provider || '-'],
-    [translateRealtimeErrorText('column_model', t, language), row.model || '-'],
-  ];
-  return fields.map(([label, value]) => `${label}: ${maskSensitiveText(String(value ?? '-'))}`).join('\n');
-};
 
 const ACCOUNT_QUOTA_RENDER_HELPERS: QuotaRenderHelpers = {
   styles: {
@@ -2254,915 +697,6 @@ const buildAccountQuotaEntriesByAccount = (
     } satisfies AccountQuotaEntry)),
   ])
 );
-
-const buildRealtimeLogPageRows = (
-  rows: MonitoringEventRow[],
-  page: number,
-  pageSize: number
-): { total: number; rows: RealtimeLogRow[] } => {
-  const candidateRows = rows.length > REALTIME_LOG_ENRICH_LIMIT
-    ? rows.slice(0, REALTIME_LOG_ENRICH_LIMIT)
-    : rows;
-  const metricsByStream = new Map<string, { total: number; success: number; pattern: boolean[] }>();
-  const normalizedPage = Math.max(1, page);
-  const pageStart = (normalizedPage - 1) * pageSize;
-  const pageEnd = Math.min(pageStart + pageSize, candidateRows.length);
-  const enriched = new Array<RealtimeLogRow>(Math.max(pageEnd - pageStart, 0));
-
-  for (let index = candidateRows.length - 1; index >= 0; index -= 1) {
-    const row = candidateRows[index];
-    const streamKey = [row.account, row.provider, row.model, row.modelAlias, row.channel].join('::');
-    const previous = metricsByStream.get(streamKey) ?? { total: 0, success: 0, pattern: [] };
-    const nextPattern = [...previous.pattern, !row.failed].slice(-10);
-    const next = {
-      total: previous.total + (row.statsIncluded ? 1 : 0),
-      success: previous.success + (row.statsIncluded && !row.failed ? 1 : 0),
-      pattern: nextPattern,
-    };
-    metricsByStream.set(streamKey, next);
-
-    if (index >= pageStart && index < pageEnd) {
-      let recentSuccessCount = 0;
-      nextPattern.forEach((item) => {
-        if (item) recentSuccessCount += 1;
-      });
-      enriched[index - pageStart] = {
-        ...row,
-        streamKey,
-        diagnosticText: buildRealtimeDiagnosticText(row),
-        errorCategoryKey: resolveRealtimeErrorCategoryKey(row),
-        errorSummary: buildRealtimeErrorSummary(row),
-        requestCount: next.total,
-        successRate: next.total > 0 ? next.success / next.total : 1,
-        recentPattern: nextPattern,
-        recentSuccessCount,
-        recentFailureCount: nextPattern.length - recentSuccessCount,
-      };
-    }
-  }
-
-  return { total: candidateRows.length, rows: enriched };
-};
-
-const getClientPaginationRange = (page: number, pageSize: number, total: number, visibleCount: number) => {
-  const normalizedPage = Math.max(1, page);
-  const totalPages = total > 0 ? Math.ceil(total / pageSize) : 0;
-  const from = total > 0 && visibleCount > 0 ? (normalizedPage - 1) * pageSize + 1 : 0;
-  return {
-    page: normalizedPage,
-    total,
-    totalPages,
-    from,
-    to: visibleCount > 0 ? Math.min(total, from + visibleCount - 1) : 0,
-    hasPrevious: normalizedPage > 1,
-    hasNext: normalizedPage < totalPages,
-  };
-};
-
-function UsageTrendHeader({
-  range,
-  totalCalls,
-  apiKeyFilter,
-  apiKeyOptions,
-  onRangeChange,
-  onApiKeyFilterChange,
-  onHide,
-  t,
-}: {
-  range: MonitoringTimeRange;
-  totalCalls: number;
-  apiKeyFilter: string;
-  apiKeyOptions: Array<{ value: string; label: string }>;
-  onRangeChange: (range: MonitoringTimeRange) => void;
-  onApiKeyFilterChange: (value: string) => void;
-  onHide: () => void;
-  t: TFunction;
-}) {
-  return (
-    <div className={styles.usageTrendHeader}>
-      <div className={styles.usageTrendCopy}>
-        <h2>{t('monitoring.usage_stats_title')}</h2>
-        <p>{t('monitoring.usage_stats_desc', { value: formatCompactNumber(totalCalls) })}</p>
-      </div>
-      <button type="button" className={`${styles.rankingMetricButton} ${styles.usageTrendHideButton} ${styles.mobileHeaderHideButton}`} onClick={onHide}>
-        {t('monitoring.hide_analysis')}
-      </button>
-      <div className={styles.usageTrendActions}>
-        <div className={`${styles.rankingMetricSwitch} ${styles.timeRangeControl}`}>
-          {TIME_RANGE_OPTIONS.map((option) => (
-            <button
-              key={option.value}
-              type="button"
-              className={`${styles.rankingMetricButton} ${styles.timeRangeButton} ${range === option.value ? styles.rankingMetricButtonActive : ''}`}
-              onClick={() => onRangeChange(option.value)}
-            >
-              {t(option.labelKey)}
-            </button>
-          ))}
-        </div>
-        <Select
-          className={styles.usageTrendApiKeySelect}
-          value={apiKeyFilter}
-          options={apiKeyOptions}
-          onChange={onApiKeyFilterChange}
-          ariaLabel={t('monitoring.filter_usage_trend_api_key')}
-          fullWidth={false}
-        />
-        <button type="button" className={`${styles.rankingMetricButton} ${styles.usageTrendHideButton}`} onClick={onHide}>
-          {t('monitoring.hide_analysis')}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function TopUsageStats({ cards }: { cards: UsageMetricCard[] }) {
-  return (
-    <section className={styles.usageStatsGrid} aria-label="Usage statistics">
-      {cards.map((card) => (
-        <Card key={card.key} className={`${styles.usageStatsCard} ${card.key === 'tokens' ? styles.usageStatsCardTokens : ''}`}>
-          <div className={styles.usageStatsCardHeader}>
-            <span className={`${styles.usageStatsIcon} ${styles[`usageStatsIcon${card.accent}`]}`} aria-hidden="true" />
-            <strong>{card.title}</strong>
-          </div>
-          <div className={styles.usageStatsBody}>
-            <span>{card.label}</span>
-            <strong>{card.value}</strong>
-          </div>
-          <div className={styles.usageStatsFooter}>
-            {card.footer.map((item) => (
-              <div key={item.label}>
-                <span>{item.label}</span>
-                <strong>{item.value}</strong>
-              </div>
-            ))}
-          </div>
-        </Card>
-      ))}
-    </section>
-  );
-}
-
-function UsageTrendPanel({
-  points,
-  hasPrices,
-  emptyText,
-  t,
-}: {
-  points: TrendPoint[];
-  hasPrices: boolean;
-  emptyText: string;
-  t: TFunction;
-}) {
-  const chartPoints = points.slice(-30);
-  const [hoveredIndex, setHoveredIndex] = useState<number | null>(null);
-  const svgRef = useRef<SVGSVGElement | null>(null);
-  const chartViewBoxHeight = 310;
-  const [chartViewBoxWidth, setChartViewBoxWidth] = useState(700);
-  const rightLabelX = chartViewBoxWidth - 8;
-  const plot = {
-    left: 36,
-    top: 38,
-    right: rightLabelX - 144,
-    costAxis: rightLabelX - 88,
-    tokenAxis: rightLabelX - 42,
-    tokenLabel: rightLabelX,
-    bottom: 278,
-  };
-  const requestMax = Math.max(...chartPoints.map((point) => point.requests), 0);
-  const tokenMax = Math.max(...chartPoints.map((point) => point.tokens), 0);
-  const costMax = Math.max(...chartPoints.map((point) => point.cost), 0);
-  const requestAxisMax = Math.max(10, Math.ceil(requestMax * 1.1));
-  const tokenAxisMax = Math.max(1000, Math.ceil(tokenMax * 1.1));
-  const costAxisMax = Math.max(0.1, costMax * 1.1);
-  const series = [
-    {
-      key: 'tokens',
-      label: t('monitoring.ranking_metric_tokens'),
-      color: '#7c3aed',
-      axis: 'tokens',
-      getValue: (point: TrendPoint) => point.tokens,
-      format: (value: number) => formatCompactNumber(value),
-    },
-    {
-      key: 'requests',
-      label: t('monitoring.ranking_metric_requests'),
-      color: '#2563eb',
-      axis: 'requests',
-      getValue: (point: TrendPoint) => point.requests,
-      format: (value: number) => formatCompactNumber(value),
-    },
-    {
-      key: 'cost',
-      label: t('monitoring.ranking_metric_cost'),
-      color: '#047857',
-      axis: 'cost',
-      getValue: (point: TrendPoint) => point.cost,
-      format: (value: number) => (hasPrices ? formatUsd(value) : '--'),
-    },
-  ] as const;
-  const visibleSeries = hasPrices ? series : series.filter((item) => item.key !== 'cost');
-  const totals = chartPoints.reduce(
-    (sum, point) => ({
-      requests: sum.requests + point.requests,
-      failures: sum.failures + point.failures,
-      tokens: sum.tokens + point.tokens,
-      cost: sum.cost + point.cost,
-    }),
-    { requests: 0, failures: 0, tokens: 0, cost: 0 }
-  );
-  const peakTokenPoint = chartPoints.reduce<TrendPoint | null>(
-    (peak, point) => (!peak || point.tokens > peak.tokens ? point : peak),
-    null
-  );
-  const summaryItems = [
-    { key: 'requests', label: t('monitoring.total_requests_label'), value: formatCompactNumber(totals.requests), color: '#2563eb' },
-    { key: 'tokens', label: t('monitoring.total_tokens_label'), value: formatCompactNumber(totals.tokens), color: '#7c3aed' },
-    ...(hasPrices ? [{ key: 'cost', label: t('monitoring.total_cost_label'), value: formatUsd(totals.cost), color: '#059669' }] : []),
-    { key: 'peak', label: t('monitoring.peak_period'), value: peakTokenPoint?.label ?? '--', color: '#f97316' },
-  ];
-  const trendMinutes = Math.max(chartPoints.length * 60, 1);
-  const headerStats = [
-    { key: 'rpm', label: 'RPM', value: (totals.requests / trendMinutes).toFixed(2) },
-    { key: 'tpm', label: 'TPM', value: formatCompactNumber(totals.tokens / trendMinutes) },
-    { key: 'errorRate', label: t('monitoring.error_rate'), value: formatPercent(totals.requests > 0 ? totals.failures / totals.requests : 0) },
-  ];
-  const axisTicks = [0, 0.25, 0.5, 0.75, 1];
-  const getAxisMax = (axis: typeof series[number]['axis']) => {
-    if (axis === 'tokens') return tokenAxisMax;
-    if (axis === 'cost') return costAxisMax;
-    return requestAxisMax;
-  };
-  const getX = (index: number) => chartPoints.length <= 1
-    ? (plot.left + plot.right) / 2
-    : plot.left + (index / (chartPoints.length - 1)) * (plot.right - plot.left);
-  const getY = (value: number, axis: typeof series[number]['axis']) => {
-    const max = getAxisMax(axis);
-    return plot.bottom - Math.max(Math.min(value / max, 1), 0) * (plot.bottom - plot.top);
-  };
-  const buildPath = (item: typeof series[number]) => {
-    const coords = chartPoints.map((point, index) => ({
-      x: getX(index),
-      y: getY(item.getValue(point), item.axis),
-    }));
-    if (coords.length === 0) return '';
-    if (coords.length === 1) return `M ${coords[0].x} ${coords[0].y}`;
-    return coords.slice(1).reduce((path, point, index) => {
-      const previous = coords[index];
-      const midX = (previous.x + point.x) / 2;
-      return `${path} C ${midX} ${previous.y}, ${midX} ${point.y}, ${point.x} ${point.y}`;
-    }, `M ${coords[0].x} ${coords[0].y}`);
-  };
-  const buildAreaPath = (item: typeof series[number]) => {
-    const path = buildPath(item);
-    return path ? `${path} L ${getX(chartPoints.length - 1)} ${plot.bottom} L ${getX(0)} ${plot.bottom} Z` : '';
-  };
-  const labels = getChartAxisLabels(chartPoints);
-  const hoveredPoint = hoveredIndex === null ? null : chartPoints[hoveredIndex];
-  const hoveredX = hoveredIndex === null ? 0 : getX(hoveredIndex);
-  const tooltipX = Math.min(Math.max(hoveredX - 84, plot.left + 8), plot.right - 168);
-  const formatCostAxisValue = (value: number) => `$${formatCompactNumber(value)}`;
-
-  useEffect(() => {
-    const svg = svgRef.current;
-    if (!svg) return;
-
-    const updateViewBoxWidth = () => {
-      const rect = svg.getBoundingClientRect();
-      const nextWidth = Math.max(700, Math.round((rect.width / Math.max(rect.height, 1)) * chartViewBoxHeight));
-      setChartViewBoxWidth((current) => current === nextWidth ? current : nextWidth);
-    };
-
-    updateViewBoxWidth();
-    const observer = new ResizeObserver(updateViewBoxWidth);
-    observer.observe(svg);
-    return () => observer.disconnect();
-  }, [chartViewBoxHeight]);
-
-  return (
-    <Card className={`${styles.usageTrendChartCard} ${styles.usageTrendLineCard}`}>
-      <div className={styles.trendCardHeader}>
-        <div>
-          <h3>{t('monitoring.usage_trend_chart_title')}</h3>
-          <p>{t('monitoring.usage_trend_chart_desc')}</p>
-        </div>
-        <div className={styles.trendHeaderStats}>
-          {headerStats.map((item) => (
-            <div key={item.key}>
-              <span>{item.label}</span>
-              <strong>{item.value}</strong>
-            </div>
-          ))}
-        </div>
-      </div>
-      {chartPoints.length > 0 ? (
-        <div className={styles.professionalChartShell}>
-          <div className={styles.trendSummaryStrip}>
-            {summaryItems.map((item) => (
-              <div key={item.key} className={styles.trendSummaryItem} style={{ '--series-color': item.color } as CSSProperties}>
-                <span>{item.label}</span>
-                <strong>{item.value}</strong>
-              </div>
-            ))}
-          </div>
-          <svg ref={svgRef} className={styles.usageTrendSvg} viewBox={`0 0 ${chartViewBoxWidth} ${chartViewBoxHeight}`} role="img" aria-label={t('monitoring.usage_trend_chart_aria')}>
-            <defs>
-              <linearGradient id="usageTrendTokensFill" x1="0" x2="0" y1="0" y2="1">
-                <stop offset="5%" stopColor="#7c3aed" stopOpacity="0.24" />
-                <stop offset="95%" stopColor="#7c3aed" stopOpacity="0" />
-              </linearGradient>
-              <linearGradient id="usageTrendRequestsFill" x1="0" x2="0" y1="0" y2="1">
-                <stop offset="5%" stopColor="#2563eb" stopOpacity="0.16" />
-                <stop offset="95%" stopColor="#2563eb" stopOpacity="0" />
-              </linearGradient>
-              <linearGradient id="usageTrendCostFill" x1="0" x2="0" y1="0" y2="1">
-                <stop offset="5%" stopColor="#047857" stopOpacity="0.24" />
-                <stop offset="95%" stopColor="#047857" stopOpacity="0" />
-              </linearGradient>
-            </defs>
-            {axisTicks.map((tick) => {
-              const y = plot.bottom - tick * (plot.bottom - plot.top);
-              return (
-                <g key={tick}>
-                  {tick > 0 ? <line className={styles.chartGridLine} x1={plot.left} x2={plot.costAxis} y1={y} y2={y} /> : null}
-                  <text className={`${styles.chartAxisLabel} ${styles.chartAxisLabelRequests}`} x={plot.left - 12} y={y + 4}>
-                    {formatCompactNumber(requestAxisMax * tick)}
-                  </text>
-                  {hasPrices ? (
-                    <text className={`${styles.chartAxisLabel} ${styles.chartAxisLabelCost}`} x={plot.costAxis - 12} y={y + 4}>
-                      {formatCostAxisValue(costAxisMax * tick)}
-                    </text>
-                  ) : null}
-                  <text className={`${styles.chartAxisLabel} ${styles.chartAxisLabelTokens}`} x={plot.tokenLabel} y={y + 4}>
-                    {formatCompactNumber(tokenAxisMax * tick)}
-                  </text>
-                </g>
-              );
-            })}
-            <line className={styles.chartAxisBase} x1={plot.left} x2={plot.costAxis} y1={plot.bottom} y2={plot.bottom} />
-            <line className={styles.chartYAxisRequests} x1={plot.left} x2={plot.left} y1={plot.top} y2={plot.bottom} />
-            {hasPrices ? <line className={styles.chartYAxisCost} x1={plot.costAxis} x2={plot.costAxis} y1={plot.top} y2={plot.bottom} /> : null}
-            <line className={styles.chartYAxisTokens} x1={plot.tokenAxis} x2={plot.tokenAxis} y1={plot.top} y2={plot.bottom} />
-            {visibleSeries.map((item) => {
-              const path = buildPath(item);
-              const area = buildAreaPath(item);
-              return (
-                <g key={item.key}>
-                  {area ? <path className={styles.trendAreaFill} d={area} fill={`url(#usageTrend${item.key[0].toUpperCase()}${item.key.slice(1)}Fill)`} /> : null}
-                  <path className={styles.trendSeriesLine} d={path} stroke={item.color} />
-                </g>
-              );
-            })}
-            {labels.map((item) => (
-              <g key={item.key}>
-                <line className={styles.chartXAxisTick} x1={getX(item.index)} x2={getX(item.index)} y1={plot.bottom} y2={plot.bottom + 7} />
-                <text className={styles.chartXAxisLabel} x={getX(item.index)} y="300">{item.label}</text>
-              </g>
-            ))}
-            {chartPoints.map((point, index) => {
-              const x = getX(index);
-              const isHovered = hoveredIndex === index;
-              return (
-                <g
-                  key={point.key}
-                  className={styles.trendHoverTarget}
-                  onMouseEnter={() => setHoveredIndex(index)}
-                  onMouseLeave={() => setHoveredIndex(null)}
-                  onFocus={() => setHoveredIndex(index)}
-                  onBlur={() => setHoveredIndex(null)}
-                  tabIndex={0}
-                >
-                  <rect x={Math.max(plot.left, x - 14)} y={plot.top - 10} width="28" height={plot.bottom - plot.top + 26} fill="transparent" />
-                  {isHovered ? <line className={styles.trendHoverGuide} x1={x} x2={x} y1={plot.top} y2={plot.bottom} /> : null}
-                  {isHovered ? visibleSeries.map((item) => (
-                    <circle
-                      key={item.key}
-                      className={styles.trendSeriesDot}
-                      cx={x}
-                      cy={getY(item.getValue(point), item.axis)}
-                      r={4.5}
-                      stroke={item.color}
-                    />
-                  )) : null}
-                </g>
-              );
-            })}
-            {hoveredPoint ? (
-              <g className={styles.trendTooltipLayer}>
-                <rect x={tooltipX} y="82" width="168" height={hasPrices ? 118 : 92} rx="12" />
-                <text className={styles.trendTooltipTitle} x={tooltipX + 16} y="108">{hoveredPoint.label}</text>
-                {visibleSeries.map((item, index) => (
-                  <text key={item.key} className={styles.trendTooltipMetric} x={tooltipX + 16} y={136 + index * 25} fill={item.color}>
-                    {`${item.label}：${item.format(item.getValue(hoveredPoint))}`}
-                  </text>
-                ))}
-              </g>
-            ) : null}
-          </svg>
-          <div className={styles.trendChartLegend}>
-            {visibleSeries.map((item) => (
-              <span key={item.key} style={{ '--series-color': item.color } as CSSProperties}>
-                {item.label}
-              </span>
-            ))}
-          </div>
-        </div>
-      ) : (
-        <div className={styles.emptyBlockSmall}>{emptyText}</div>
-      )}
-    </Card>
-  );
-}
-
-function TokenDistributionPanel({
-  points,
-  emptyText,
-  hasPrices,
-  t,
-}: {
-  points: TokenDistributionPoint[];
-  emptyText: string;
-  hasPrices: boolean;
-  t: TFunction;
-}) {
-  const totals = points.reduce(
-    (sum, point) => ({
-      requests: sum.requests + point.requests,
-      totalTokens: sum.totalTokens + point.totalTokens,
-      inputTokens: sum.inputTokens + point.inputTokens,
-      outputTokens: sum.outputTokens + point.outputTokens,
-      reasoningTokens: sum.reasoningTokens + point.reasoningTokens,
-      cachedTokens: sum.cachedTokens + point.cachedTokens,
-      totalCost: sum.totalCost + point.totalCost,
-    }),
-    { requests: 0, totalTokens: 0, inputTokens: 0, outputTokens: 0, reasoningTokens: 0, cachedTokens: 0, totalCost: 0 }
-  );
-  const tokenMinutes = Math.max(points.length * 60, 1);
-  const rpm = totals.requests / tokenMinutes;
-  const tpm = totals.totalTokens / tokenMinutes;
-  const rows = [
-    { key: 'rpm', label: 'RPM', value: rpm, displayValue: rpm.toFixed(2), base: 0, accent: 'Purple', showShare: false },
-    { key: 'tpm', label: 'TPM', value: tpm, displayValue: formatCompactNumber(tpm), base: 0, accent: 'Blue', showShare: false },
-    { key: 'requests', label: t('monitoring.request_count'), value: totals.requests, displayValue: formatCompactNumber(totals.requests), base: 0, accent: 'Cyan', showShare: false },
-    { key: 'total', label: t('monitoring.total_tokens_label'), value: totals.totalTokens, displayValue: formatCompactNumber(totals.totalTokens), base: 0, accent: 'Green', showShare: false },
-    { key: 'input', label: t('monitoring.token_metric_input'), value: totals.inputTokens, displayValue: formatCompactNumber(totals.inputTokens), base: totals.totalTokens, accent: 'Amber', showShare: true },
-    { key: 'output', label: t('monitoring.token_metric_output'), value: totals.outputTokens, displayValue: formatCompactNumber(totals.outputTokens), base: totals.totalTokens, accent: 'Rose', showShare: true },
-    { key: 'reasoning', label: t('monitoring.token_metric_reasoning'), value: totals.reasoningTokens, displayValue: formatCompactNumber(totals.reasoningTokens), base: totals.totalTokens, accent: 'Indigo', showShare: true },
-    { key: 'cached', label: t('monitoring.token_metric_cached'), value: totals.cachedTokens, displayValue: formatCompactNumber(totals.cachedTokens), base: totals.totalTokens, accent: 'Slate', showShare: true },
-  ];
-  const hasData = rows.some((row) => row.value > 0);
-
-  return (
-    <Card className={`${styles.usageTrendChartCard} ${styles.tokenDistributionCard}`}>
-      <div className={`${styles.trendCardHeader} ${styles.tokenDistributionHeader}`}>
-        <div>
-          <h3>{t('monitoring.token_stats_title')}</h3>
-          <p>{t('monitoring.token_stats_desc')}</p>
-        </div>
-        <div className={styles.tokenCostBadge}>
-          <span>{t('monitoring.token_cost')}</span>
-          <strong>{hasPrices ? formatUsd(totals.totalCost) : '--'}</strong>
-        </div>
-      </div>
-      {hasData ? (
-        <div className={styles.tokenStatCardList}>
-          {rows.map((row) => {
-            const share = row.base > 0 ? row.value / row.base : 0;
-            return (
-              <div key={row.key} className={`${styles.tokenStatCard} ${styles[`tokenStatCard${row.accent}`]}`}>
-                <div className={styles.tokenStatCardHeader}>
-                  <span>{row.label}</span>
-                  {row.showShare ? <strong>{formatPercent(share)}</strong> : null}
-                </div>
-                <div className={styles.tokenStatCardValue}>{row.displayValue}</div>
-                {row.showShare ? (
-                  <div className={styles.tokenStatProgressTrack} aria-hidden="true">
-                    <span style={{ '--token-stat-width': `${Math.min(Math.max(share * 100, row.value > 0 ? 1 : 0), 100)}%` } as CSSProperties} />
-                  </div>
-                ) : null}
-              </div>
-            );
-          })}
-        </div>
-      ) : (
-        <div className={styles.emptyBlockSmall}>{emptyText}</div>
-      )}
-    </Card>
-  );
-}
-
-function ModelStatsPanel({
-  title,
-  subtitle,
-  rows,
-  metric,
-  metricTotal,
-  onMetricChange,
-  emptyText,
-  hasPrices,
-  t,
-}: {
-  title: string;
-  subtitle: string;
-  rows: MonitoringAccountRow[];
-  metric: RankingMetric;
-  metricTotal: number;
-  onMetricChange: (metric: RankingMetric) => void;
-  emptyText: string;
-  hasPrices: boolean;
-  t: TFunction;
-}) {
-  const shareBase = metricTotal > 0 ? metricTotal : rows.reduce((sum, row) => sum + getRankingMetricValue(row, metric), 0);
-  const shareModeLabel = t('monitoring.share_by_metric', { metric: getRankingMetricLabel(metric, t) });
-  const totalShareValue = formatRankingMetricValue(shareBase, metric, hasPrices);
-  const legendRows = rows.slice(0, 5);
-  const donutTooltipRows = rows
-    .map((row) => {
-      const value = getRankingMetricValue(row, metric);
-      return {
-        row,
-        value,
-        share: shareBase > 0 ? value / shareBase : 0,
-      };
-    })
-    .filter((item) => item.value > 0);
-  const donutStops = legendRows.reduce(
-    (state, row, index) => {
-      const value = getRankingMetricValue(row, metric);
-      const share = shareBase > 0 ? (value / shareBase) * 100 : 0;
-      const end = state.offset + share;
-      state.parts.push(`${getDonutColor(index)} ${state.offset}% ${end}%`);
-      state.offset = end;
-      return state;
-    },
-    { offset: 0, parts: [] as string[] }
-  );
-  const donutBackground = donutStops.parts.length > 0
-    ? `conic-gradient(${donutStops.parts.join(', ')}, color-mix(in srgb, var(--monitor-line) 58%, transparent) ${donutStops.offset}% 100%)`
-    : 'color-mix(in srgb, var(--monitor-line) 58%, transparent)';
-
-  return (
-    <Card className={styles.modelStatsCard}>
-      <div className={styles.rankingHeader}>
-        <div>
-          <h3>{title}</h3>
-          <p>{subtitle}</p>
-        </div>
-        <RankingMetricSwitch value={metric} onChange={onMetricChange} disabledCost={!hasPrices} t={t} />
-      </div>
-      {rows.length > 0 ? (
-        <div className={styles.modelStatsLayout}>
-          <div className={styles.modelStatsList}>
-            {rows.map((row) => {
-              const rowValue = getRankingMetricValue(row, metric);
-              const value = shareBase > 0 ? rowValue / shareBase : 0;
-              return (
-                <div key={row.id} className={styles.modelStatsRow}>
-                  <div className={styles.modelStatsMain}>
-                    <div className={styles.modelStatsTitleLine}>
-                      <strong>{row.account}</strong>
-                      <span>{formatPercent(value)}</span>
-                    </div>
-                    <div className={styles.modelStatsMetaLine}>
-                      <span>{`${formatCompactNumber(row.totalCalls)} ${t('monitoring.ranking_metric_requests')}`}</span>
-                      <span>{`${formatCompactNumber(row.totalTokens)} ${t('monitoring.ranking_metric_tokens')}`}</span>
-                      <span>{`${formatCompactNumber(row.failureCalls)} ${t('monitoring.errors_label')}`}</span>
-                      <span>{hasPrices ? formatUsd(row.totalCost) : '--'}</span>
-                    </div>
-                    <span
-                      className={styles.modelStatsBar}
-                      style={{ '--ranking-width': getProgressWidth(value) } as CSSProperties}
-                      aria-hidden="true"
-                    />
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-          <aside className={styles.modelSharePanel}>
-            <div className={styles.modelShareHeader}>
-              <strong>{t('monitoring.model_share_title')}</strong>
-              <span>{shareModeLabel}</span>
-            </div>
-            <div className={styles.donutChart} style={{ '--donut-bg': donutBackground } as CSSProperties}>
-              <div className={styles.donutCenter}>
-                <span>{getRankingSummaryLabel(metric, t)}</span>
-                <strong>{totalShareValue}</strong>
-              </div>
-              <div className={styles.donutTooltip}>
-                <strong>{t('monitoring.share_tooltip_title', { metric: shareModeLabel })}</strong>
-                <div>
-                  {donutTooltipRows.map((item, index) => (
-                    <span key={item.row.id}>
-                      <i style={{ background: getDonutColor(index) }} aria-hidden="true" />
-                      <em>{item.row.account}</em>
-                      <b>{formatPercent(item.share)}</b>
-                    </span>
-                  ))}
-                </div>
-              </div>
-            </div>
-            <div className={styles.modelLegend}>
-              {legendRows.map((row, index) => {
-                const rowValue = getRankingMetricValue(row, metric);
-                const value = shareBase > 0 ? rowValue / shareBase : 0;
-                return (
-                  <div key={row.id} className={styles.modelLegendItem} style={{ '--legend-color': getDonutColor(index) } as CSSProperties}>
-                    <span className={styles.modelLegendDot} />
-                    <span>{row.account}</span>
-                    <strong>{formatPercent(value)}</strong>
-                  </div>
-                );
-              })}
-            </div>
-          </aside>
-        </div>
-      ) : (
-        <div className={styles.emptyBlockSmall}>{emptyText}</div>
-      )}
-    </Card>
-  );
-}
-
-function ApiKeyRankingPanel({
-  title,
-  subtitle,
-  rows,
-  metric,
-  metricTotal,
-  onMetricChange,
-  emptyText,
-  hasPrices,
-  t,
-}: {
-  title: string;
-  subtitle: string;
-  rows: MonitoringAccountRow[];
-  metric: RankingMetric;
-  metricTotal: number;
-  onMetricChange: (metric: RankingMetric) => void;
-  emptyText: string;
-  hasPrices: boolean;
-  t: TFunction;
-}) {
-  const shareBase = metricTotal > 0 ? metricTotal : rows.reduce((sum, row) => sum + getRankingMetricValue(row, metric), 0);
-  const summaryLabel = getRankingSummaryLabel(metric, t);
-
-  return (
-    <Card className={styles.apiKeyRankingCard}>
-      <div className={styles.rankingHeader}>
-        <div>
-          <h3>{title}</h3>
-          <p>{subtitle}</p>
-        </div>
-        <RankingMetricSwitch value={metric} onChange={onMetricChange} disabledCost={!hasPrices} t={t} />
-      </div>
-      <div className={styles.apiKeyRankingList}>
-        {rows.length > 0 ? (
-          <div className={styles.apiKeyRankingSummary}>
-            <span>{summaryLabel}</span>
-            <strong>{formatRankingMetricValue(shareBase, metric, hasPrices)}</strong>
-            <small>{t('monitoring.api_keys_count', { count: rows.length })}</small>
-          </div>
-        ) : null}
-        {rows.length > 0 ? (
-          <div className={styles.apiKeyRankingScroll}>
-            {rows.map((row, index) => {
-              const rowValue = getRankingMetricValue(row, metric);
-              const share = shareBase > 0 ? rowValue / shareBase : 0;
-              return (
-                <div key={row.id} className={styles.apiKeyRankingRow}>
-                  <div className={styles.apiKeyRankingTopLine}>
-                    <div className={styles.apiKeyRankingName}>
-                      <span className={styles.rankingIndex}>{index + 1}</span>
-                      <strong>{row.account}</strong>
-                    </div>
-                    <span>{formatPercent(share)}</span>
-                  </div>
-                  <div className={styles.apiKeyRankingMetaLine}>
-                    <span>{`${formatCompactNumber(row.totalCalls)} ${t('monitoring.ranking_metric_requests')}`}</span>
-                    <span>{`${formatCompactNumber(row.totalTokens)} Token`}</span>
-                    <span>{`${formatCompactNumber(row.failureCalls)} ${t('monitoring.errors_label')}`}</span>
-                    <span>{hasPrices ? formatUsd(row.totalCost) : '--'}</span>
-                  </div>
-                  <span
-                    className={styles.apiKeyRankingBar}
-                    style={{ '--ranking-width': getProgressWidth(share) } as CSSProperties}
-                    aria-hidden="true"
-                  />
-                </div>
-              );
-            })}
-          </div>
-        ) : (
-          <div className={styles.emptyBlockSmall}>{emptyText}</div>
-        )}
-      </div>
-    </Card>
-  );
-}
-
-function MonitoringHealthStatusBar({
-  statusData,
-  locale,
-  t,
-  showRate = true,
-}: {
-  statusData: AccountStatusData;
-  locale: string;
-  t: TFunction;
-  showRate?: boolean;
-}) {
-  const [activeTooltip, setActiveTooltip] = useState<number | null>(null);
-  const [focusIndex, setFocusIndex] = useState(() => (statusData.blockDetails.length > 0 ? 0 : -1));
-  const blocksRef = useRef<HTMLDivElement | null>(null);
-  const blockButtonRefs = useRef<Array<HTMLButtonElement | null>>([]);
-  const tooltipIdPrefix = useId();
-  const blockCount = statusData.blockDetails.length;
-  const resolvedFocusIndex = blockCount === 0 ? -1 : focusIndex >= 0 && focusIndex < blockCount ? focusIndex : 0;
-  const resolvedActiveTooltip = activeTooltip !== null && activeTooltip >= 0 && activeTooltip < blockCount ? activeTooltip : null;
-  const hasData = statusData.totalSuccess + statusData.totalFailure > 0;
-  const rateClassName = !hasData
-    ? ''
-    : statusData.successRate >= 90
-      ? styles.monitoringStatusRateHigh
-      : statusData.successRate >= 50
-        ? styles.monitoringStatusRateMedium
-        : styles.monitoringStatusRateLow;
-
-  useEffect(() => {
-    if (resolvedActiveTooltip === null) return;
-    const handlePointerDown = (event: PointerEvent) => {
-      if (blocksRef.current && !blocksRef.current.contains(event.target as Node)) {
-        setActiveTooltip(null);
-      }
-    };
-    document.addEventListener('pointerdown', handlePointerDown);
-    return () => document.removeEventListener('pointerdown', handlePointerDown);
-  }, [resolvedActiveTooltip]);
-
-  const handlePointerEnter = useCallback((event: React.PointerEvent, index: number) => {
-    if (event.pointerType === 'mouse') {
-      setActiveTooltip(index);
-    }
-  }, []);
-
-  const handlePointerLeave = useCallback((event: React.PointerEvent) => {
-    if (event.pointerType === 'mouse' && (!blocksRef.current || !blocksRef.current.contains(document.activeElement))) {
-      setActiveTooltip(null);
-    }
-  }, []);
-
-  const handlePointerDown = useCallback((event: React.PointerEvent, index: number) => {
-    if (event.pointerType === 'touch') {
-      event.preventDefault();
-      setFocusIndex(index);
-      setActiveTooltip((previous) => (previous === index ? null : index));
-    }
-  }, []);
-
-  const focusBlock = useCallback((index: number) => {
-    blockButtonRefs.current[index]?.focus();
-    setFocusIndex(index);
-    setActiveTooltip(index);
-  }, []);
-
-  const handleFocus = useCallback((index: number) => {
-    setFocusIndex(index);
-    setActiveTooltip(index);
-  }, []);
-
-  const handleBlur = useCallback((event: React.FocusEvent<HTMLButtonElement>) => {
-    if (blocksRef.current?.contains(event.relatedTarget as Node | null)) {
-      return;
-    }
-    setActiveTooltip(null);
-  }, []);
-
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLButtonElement>, index: number) => {
-      if (event.key === 'Escape') {
-        setActiveTooltip(null);
-        return;
-      }
-
-      const nextIndex = getNextAccountStatusBlockIndex(index, event.key, blockCount);
-      if (nextIndex === null) return;
-
-      event.preventDefault();
-      focusBlock(nextIndex);
-    },
-    [blockCount, focusBlock]
-  );
-
-  const getTooltipPositionClassName = (index: number, total: number) => {
-    if (index <= 2) return styles.monitoringStatusTooltipLeft;
-    if (index >= total - 3) return styles.monitoringStatusTooltipRight;
-    return '';
-  };
-
-  const renderTooltip = (detail: AccountStatusBlockDetail, index: number, tooltipId: string) => {
-    const total = detail.success + detail.failure;
-    const timeRange = formatStatusWindowLabel(detail.startTime, detail.endTime, locale);
-
-    return (
-      <div
-        id={tooltipId}
-        role="tooltip"
-        className={[
-          styles.monitoringStatusTooltip,
-          getTooltipPositionClassName(index, statusData.blockDetails.length),
-        ]
-          .filter(Boolean)
-          .join(' ')}
-      >
-        <span className={styles.monitoringTooltipTime}>{timeRange}</span>
-        {total > 0 ? (
-          <span className={styles.monitoringTooltipStats}>
-            <span className={styles.monitoringTooltipSuccess}>{t('status_bar.success_short')} {detail.success}</span>
-            <span className={styles.monitoringTooltipFailure}>{t('status_bar.failure_short')} {detail.failure}</span>
-            <span className={styles.monitoringTooltipRate}>({(detail.rate * 100).toFixed(1)}%)</span>
-          </span>
-        ) : (
-          <span className={styles.monitoringTooltipStats}>{t('status_bar.no_requests')}</span>
-        )}
-      </div>
-    );
-  };
-
-  return (
-    <div className={styles.monitoringStatusBar}>
-      <div
-        className={styles.monitoringStatusBlocks}
-        ref={blocksRef}
-        role="group"
-        aria-label={t('monitoring.account_health_status')}
-      >
-        {statusData.blockDetails.map((detail, index) => {
-          const isIdle = detail.rate === -1;
-          const isActive = resolvedActiveTooltip === index;
-          const timeRangeLabel = formatStatusWindowLabel(detail.startTime, detail.endTime, locale);
-          const tooltipId = `${tooltipIdPrefix}-monitoring-status-tooltip-${index}`;
-          const ariaLabel = buildAccountStatusBlockAriaLabel({
-            detail,
-            timeRangeLabel,
-            successRateValue: formatStatusRate(Math.max(0, detail.rate * 100)),
-            copy: {
-              successLabel: t('stats.success'),
-              failureLabel: t('stats.failure'),
-              noRequestsLabel: t('status_bar.no_requests'),
-              successRateLabel: t('monitoring.success_rate'),
-            },
-          });
-
-          return (
-            <div
-              key={`${detail.startTime}-${detail.endTime}`}
-              className={[styles.monitoringStatusBlockWrapper, isActive ? styles.monitoringStatusBlockActive : '']
-                .filter(Boolean)
-                .join(' ')}
-            >
-              <button
-                ref={(node) => {
-                  blockButtonRefs.current[index] = node;
-                }}
-                type="button"
-                className={styles.monitoringStatusBlockButton}
-                tabIndex={resolvedFocusIndex === index ? 0 : -1}
-                aria-label={ariaLabel}
-                aria-describedby={isActive ? tooltipId : undefined}
-                onFocus={() => handleFocus(index)}
-                onBlur={handleBlur}
-                onKeyDown={(event) => handleKeyDown(event, index)}
-                onPointerEnter={(event) => handlePointerEnter(event, index)}
-                onPointerLeave={handlePointerLeave}
-                onPointerDown={(event) => handlePointerDown(event, index)}
-              >
-                <div
-                  aria-hidden="true"
-                  className={[styles.monitoringStatusBlock, isIdle ? styles.monitoringStatusBlockIdle : '']
-                    .filter(Boolean)
-                    .join(' ')}
-                  style={isIdle ? undefined : { backgroundColor: getAccountStatusColor(detail.rate) }}
-                />
-              </button>
-              {isActive ? renderTooltip(detail, index, tooltipId) : null}
-            </div>
-          );
-        })}
-      </div>
-      {showRate ? (
-        <span
-          className={[styles.monitoringStatusRate, rateClassName, !hasData ? styles.monitoringStatusRatePlaceholder : '']
-            .filter(Boolean)
-            .join(' ')}
-        >
-          {hasData ? formatStatusRate(statusData.successRate) : '--'}
-        </span>
-      ) : null}
-    </div>
-  );
-}
 
 const getSuccessRateClassName = (rate: number) =>
   rate >= 0.95 ? styles.goodText : rate >= 0.85 ? styles.warnText : styles.badText;
